\name{amarillo}
\alias{amarillo}
\docType{data}
\title{Amarillo Soils}
\description{This sample dataset contains laboratory and published soil survey (i.e. generalized) data on the Amarillo soil series.}
\usage{data(amarillo)}
\format{
  The format is:
List of 2
 $ lab:'data.frame':	323 obs. of  40 variables:
  ..$ ID                           : int [1:323] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ user_pedon_id                : chr [1:323] "53TX305059" "53TX305059" "53TX305059" "53TX305059" ...
  ..$ user_site_id                 : chr [1:323] "53TX305059" "53TX305059" "53TX305059" "53TX305059" ...
  ..$ horizontal_datum_name        : chr [1:323] "" "" "" "" ...
  ..$ latitude_direction           : chr [1:323] "north" "north" "north" "north" ...
  ..$ latitude_degrees             : int [1:323] 33 33 33 33 33 33 33 33 33 33 ...
  ..$ latitude_minutes             : int [1:323] 10 10 10 10 10 10 11 11 11 11 ...
  ..$ latitude_seconds             : num [1:323] 30 30 30 30 30 30 5 5 5 5 ...
  ..$ longitude_direction          : chr [1:323] "west" "west" "west" "west" ...
  ..$ longitude_degrees            : int [1:323] 101 101 101 101 101 101 101 101 101 101 ...
  ..$ longitude_minutes            : int [1:323] 47 47 47 47 47 47 46 46 46 46 ...
  ..$ longitude_seconds            : num [1:323] 40 40 40 40 40 40 48 48 48 48 ...
  ..$ latitude_std_decimal_degrees : num [1:323] 33.2 33.2 33.2 33.2 33.2 ...
  ..$ longitude_std_decimal_degrees: num [1:323] -102 -102 -102 -102 -102 ...
  ..$ taxon_name                   : chr [1:323] "Amarillo" "Amarillo" "Amarillo" "Amarillo" ...
  ..$ class_type                   : chr [1:323] "correlated" "correlated" "correlated" "correlated" ...
  ..$ natural_key                  : chr [1:323] "40A34174" "40A34175" "40A34176" "40A34177" ...
  ..$ hzn_desgn                    : chr [1:323] "Ap" "B1" "Bt1" "Bt2" ...
  ..$ hzn_top                      : int [1:323] 0 18 51 86 137 190 0 20 51 76 ...
  ..$ hzn_bot                      : int [1:323] 18 51 86 137 190 236 20 51 76 102 ...
  ..$ CEC                          : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ ECEC                         : logi [1:323] NA NA NA NA NA NA ...
  ..$ ph_h2o                       : num [1:323] 7.5 7.2 7.5 8 8.4 8.4 NA NA NA NA ...
  ..$ ph_cacl2                     : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ caco3                        : int [1:323] NA NA NA 8 53 38 NA NA NA NA ...
  ..$ c_gypl2                      : int [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ EC                           : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ SAR                          : int [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ Total.Clay                   : num [1:323] 17.9 19.5 25.2 29.2 37.8 32.6 10.8 26.8 25.5 30.8 ...
  ..$ Total.Silt                   : num [1:323] 9.4 9.4 14.7 18.7 26.8 25.5 10.5 13.6 16.6 15.3 ...
  ..$ Total.Sand                   : num [1:323] 72.7 71.1 60.1 52.1 35.4 41.9 78.7 59.6 57.9 53.9 ...
  ..$ dB.33.bar                    : num [1:323] NA NA NA NA NA NA NA NA 1.2 NA ...
  ..$ dB.15.bar                    : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ LE                           : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ Water.Ret..2mm.33.bar        : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ Water.Ret.clod.33.bar        : num [1:323] NA NA NA NA NA NA NA NA 36.6 NA ...
  ..$ Water.Ret..2mm.15.bar        : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ OC_lab                       : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ OM                           : num [1:323] NA NA NA NA NA NA NA NA NA NA ...
  ..$ model_desg                   : chr [1:323] "Ap" "B" "Bt" "Bt" ...
 $ dmu:'data.frame':	383 obs. of  88 variables:
  ..$ ID             : int [1:383] 8177 8178 8179 8180 8181 8182 8183 8184 8185 8186 ...
  ..$ State          : chr [1:383] "NM" "NM" "NM" "NM" ...
  ..$ Database       : chr [1:383] "SSURGO" "SSURGO" "SSURGO" "SSURGO" ...
  ..$ Area.Symbol    : chr [1:383] "NM021" "NM021" "NM021" "NM021" ...
  ..$ Area.Name      : chr [1:383] "Harding County, New Mexico" "Harding County, New Mexico" "Harding County, New Mexico" "Harding County, New Mexico" ...
  ..$ mukey          : int [1:383] 376313 376313 376313 376379 376379 376379 376380 376380 376380 376405 ...
  ..$ Mapunit.Symbol : chr [1:383] "AM" "AM" "AM" "SR" ...
  ..$ Component.Name : chr [1:383] "Amarillo" "Amarillo" "Amarillo" "Amarillo" ...
  ..$ SIR....obsolete: chr [1:383] "TX0130" "TX0130" "TX0130" "TX0130" ...
  ..$ cokey          : int [1:383] 504646 504646 504646 504918 504918 504918 504923 504923 504923 505030 ...
  ..$ RV             : int [1:383] 85 85 85 40 40 40 40 40 40 85 ...
  ..$ Local.Phase    : chr [1:383] "" "" "" "" ...
  ..$ chkey          : int [1:383] 1113073 1113074 1113075 1113731 1113732 1113733 1113743 1113744 1113745 1114027 ...
  ..$ Designation    : chr [1:383] "H1" "H2" "H3" "H1" ...
  ..$ top            : int [1:383] 0 13 122 0 20 122 0 13 122 0 ...
  ..$ bott           : int [1:383] 13 122 165 20 122 152 13 122 152 20 ...
  ..$ CECL           : num [1:383] 5 8 8 3 8 8 3 8 8 15 ...
  ..$ CECR           : num [1:383] 10 14 14 6.5 14 14 6.5 14 14 20 ...
  ..$ S              : int [1:383] 3 3 3 3 3 3 3 3 3 3 ...
  ..$ CECH           : num [1:383] 15 20 20 10 20 20 10 20 20 25 ...
  ..$ cec_mean       : num [1:383] 10 14 14 6.5 14 14 6.5 14 14 20 ...
  ..$ ECECL          : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ ECECR          : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ S1             : int [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ ECECH          : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ PHWL           : num [1:383] 6.6 7.4 7.9 6.6 7.4 7.9 6.6 7.4 7.9 6.6 ...
  ..$ PHWR           : num [1:383] 7.2 7.9 8.2 7.2 7.9 8.2 7.2 7.9 8.2 7 ...
  ..$ PHWH           : num [1:383] 7.8 8.4 8.4 7.8 8.4 8.4 7.8 8.4 8.4 7.3 ...
  ..$ PHWM           : num [1:383] 7.2 7.9 8.15 7.2 7.9 8.15 7.2 7.9 8.15 6.95 ...
  ..$ PHCL           : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ PHCR           : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ PHCH           : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ CACOL          : int [1:383] 0 0 5 0 0 5 0 0 5 0 ...
  ..$ CACOR          : int [1:383] 3 3 7 3 3 7 3 5 7 0 ...
  ..$ CACOH          : int [1:383] 5 5 10 5 5 10 5 7 10 1 ...
  ..$ CACOM          : num [1:383] 2.5 2.5 7.5 2.5 2.5 7.5 2.5 3.5 7.5 0.5 ...
  ..$ GYPL           : int [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ GYPR           : int [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ GYPH           : int [1:383] 0 0 0 0 0 0 0 0 0 1 ...
  ..$ GYPM           : num [1:383] 0 0 0 0 0 0 0 0 0 0.5 ...
  ..$ ECL            : int [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ ECR            : num [1:383] 1 1 1 1 1 1 1 1 1 0 ...
  ..$ ECH            : num [1:383] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ ECM            : num [1:383] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ SARL           : int [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ SARR           : int [1:383] 0 0 0 1 1 1 0 0 0 0 ...
  ..$ SARH           : int [1:383] 1 1 1 2 2 2 1 1 1 2 ...
  ..$ SARM           : num [1:383] 0.5 0.5 0.5 1 1 1 0.5 0.5 0.5 1 ...
  ..$ CLAYL          : int [1:383] 10 20 20 5 20 20 5 20 20 13 ...
  ..$ CLAYR          : num [1:383] 14 27.5 27.5 10 27.5 27.5 10 27.5 27.5 21.5 ...
  ..$ CLAYH          : int [1:383] 18 35 35 15 35 35 15 35 35 30 ...
  ..$ CLAYM          : num [1:383] 14 27.5 27.5 10 27.5 27.5 10 27.5 27.5 21.5 ...
  ..$ SILL           : num [1:383] NA NA NA NA NA NA NA NA NA 30 ...
  ..$ SILR           : num [1:383] 16.4 17.4 17.4 9.2 37.8 37.8 9.2 37.8 37.8 37.1 ...
  ..$ SILH           : num [1:383] NA NA NA NA NA NA NA NA NA 45 ...
  ..$ SILM           : num [1:383] NA NA NA NA NA NA NA NA NA 37.5 ...
  ..$ SANL           : num [1:383] NA NA NA NA NA NA NA NA NA 35 ...
  ..$ SANR           : num [1:383] 69.6 55.1 55.1 80.8 34.7 34.7 80.8 34.7 34.7 41.4 ...
  ..$ SANH           : int [1:383] NA NA NA NA NA NA NA NA NA 50 ...
  ..$ SANM           : num [1:383] NA NA NA NA NA NA NA NA NA 42.5 ...
  ..$ S2             : int [1:383] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ DB3L           : num [1:383] 1.35 1.3 1.4 1.4 1.3 1.4 1.4 1.3 1.4 1.3 ...
  ..$ DB3R           : num [1:383] 1.48 1.48 1.6 1.5 1.48 1.6 1.5 1.48 1.6 1.43 ...
  ..$ DB3H           : num [1:383] 1.6 1.65 1.8 1.6 1.65 1.8 1.6 1.65 1.8 1.55 ...
  ..$ DB3M           : num [1:383] 1.47 1.47 1.6 1.5 1.47 1.6 1.5 1.47 1.6 1.42 ...
  ..$ S3             : int [1:383] 3 3 3 3 3 3 3 3 3 3 ...
  ..$ KSATL          : num [1:383] 14.11 4.23 4.23 14.11 4.23 ...
  ..$ KSATR          : num [1:383] 28.23 9.17 9.17 28.23 9.17 ...
  ..$ KSATH          : num [1:383] 42.3 14.1 14.1 42.3 14.1 ...
  ..$ KSATM          : num [1:383] 28.22 9.17 9.17 28.22 9.17 ...
  ..$ AWCL           : num [1:383] 0.11 0.14 0.1 0.06 0.14 0.1 0.06 0.14 0.1 0.12 ...
  ..$ AWCR           : num [1:383] 0.13 0.16 0.13 0.08 0.16 0.13 0.08 0.16 0.13 0.15 ...
  ..$ AWCH           : num [1:383] 0.15 0.18 0.15 0.1 0.18 0.15 0.1 0.18 0.15 0.18 ...
  ..$ AWCM           : num [1:383] 0.13 0.16 0.12 0.08 0.16 0.12 0.08 0.16 0.12 0.15 ...
  ..$ LEPL           : num [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ LEPR           : num [1:383] 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 ...
  ..$ LEPH           : num [1:383] 2.9 2.9 2.9 2.9 2.9 2.9 2.9 2.9 2.9 2.9 ...
  ..$ LEPM           : num [1:383] 1.45 1.45 1.45 1.45 1.45 1.45 1.45 1.45 1.45 1.45 ...
  ..$ OML            : num [1:383] 0.5 0.2 0.1 0.5 0.2 0.1 0.5 0.2 0.1 1 ...
  ..$ OMR            : num [1:383] 0.75 0.3 0.2 0.75 0.3 0.2 0.75 0.3 0.2 1.5 ...
  ..$ OMH            : num [1:383] 1 0.5 0.3 1 0.5 0.3 1 0.5 0.3 2 ...
  ..$ OMM            : num [1:383] 0.75 0.35 0.2 0.75 0.35 0.2 0.75 0.35 0.2 1.5 ...
  ..$ DB15L          : logi [1:383] NA NA NA NA NA NA ...
  ..$ DB15R          : num [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ DB15H          : logi [1:383] NA NA NA NA NA NA ...
  ..$ DB15M          : int [1:383] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ S4             : int [1:383] NA NA NA NA NA NA NA NA NA NA ...
  ..$ model_desg     : chr [1:383] "" "" "" "" ...
}

\source{USDA-NRCSS SSURGO database, and the KSSL database; c/o Skye Wills.}

\examples{
data(amarillo)
## maybe str(amarillo) ; plot(amarillo) ...
}
\keyword{datasets}
