% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgillicBounds.R
\name{argillic.clay.increase.depth}
\alias{argillic.clay.increase.depth}
\title{Argillic Clay Increase Depth}
\usage{
argillic.clay.increase.depth(p, clay.attr = "clay")
}
\arguments{
\item{p}{A single profile SoilProfileCollection}

\item{clay.attr}{The horizon column name containing clay data; default: \code{'clay'}}
}
\value{
A numeric vector of the increase depth, given \code{crit.clay.argillic} as the threshold and a vertical distance threshold of \code{30} cm.
}
\description{
Argillic Clay Increase Depth
}
