% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilColorIndices.R
\name{barron.torrent.redness.LAB}
\alias{barron.torrent.redness.LAB}
\title{Barron & Torrent (1986) Redness Index in LAB color space}
\usage{
barron.torrent.redness.LAB(hue, value, chroma)
}
\arguments{
\item{hue}{A character vector containing Munsell hues (e.g. "7.5YR")}

\item{value}{A numeric vector containing Munsell values}

\item{chroma}{A numeric vector containing Munsell chromas}
}
\value{
A numeric vector of horizon redness index (higher values = redder).
}
\description{
Calculate Redness Index after Barron & Torrent (1986) "Use of the Kubelka—Munk Theory to Study the Influence of Iron Oxides on Soil Colour" using Munsell colors converted to LAB. DOI: 10.1111/j.1365-2389.1986.tb00382.x. Accepts vectorized inputs for hue, value and chroma, produces vector output.
}
\references{
Barron, V. and Torrent, J. (1986), Use of the Kubelka—Munk theory to study the influence of iron oxides on soil colour. Journal of Soil Science, 37: 499-510. doi:10.1111/j.1365-2389.1986.tb00382.x
}
\author{
Andrew G. Brown
}
