% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapSoilTexture.R
\name{bootstrapSoilTexture}
\alias{bootstrapSoilTexture}
\title{Bootstrap Soil Texture Data}
\usage{
bootstrapSoilTexture(ssc, n = 100)
}
\arguments{
\item{ssc}{a \code{data.frame} object with 3 columns: `sand`, `silt`, `clay` and at least three rows of data within the range of 0-100 (percent). NA are automatically removed, but care should be taken to ensure that the sand/silt/clay values add to 100 percent. See details.}

\item{n}{number of simulated compositions. See details.}
}
\value{
a \code{list} containing:

\itemize{
 \item{samples}{\code{data.frame} of simulated sand, silt, clay values}
 \item{mean}{compositional mean}
 \item{var}{compositional variance-covariance matrix}
 \item{D.alpha}{(fitted) alpha parameters of the Dirichlet distribution}
}
}
\description{
Simulate realistic sand/silt/clay values (a composition) using the Dirichlet distribution. Parameters of the Dirichlet distribution are estimated using example data provided to the function.
}
\examples{

\donttest{
if(
requireNamespace("compositions") &
  requireNamespace("soiltexture")
) {
  
  # sample data
  data('sp6')
  depths(sp6) <- id ~ top + bottom
  
  # I still like this
  ssc <- horizons(sp6)[grep('^Bt', sp6$name), c('sand', 'silt', 'clay')]
  names(ssc) <- toupper(names(ssc))
  
  # simulate 100 samples
  s <- bootstrapSoilTexture(ssc, n = 100)
  s <- s$samples
  
  # empty soil texture triangle
  TT <- soiltexture::TT.plot(
    class.sys= "USDA-NCSS.TT",
    main= "",
    tri.sum.tst=FALSE,
    cex.lab=0.75,
    cex.axis=0.75,
    frame.bg.col='white',
    class.lab.col='black',
    lwd.axis=1.5,
    arrows.show=TRUE,
    new.mar = c(3, 0, 0, 0)
  )
  
  # add original data points
  soiltexture::TT.points(
    tri.data = s, geo = TT, col='firebrick', 
    pch = 3, cex = 0.5, lwd = 1, 
    tri.sum.tst = FALSE
  )
  
  # add simulated points
  soiltexture::TT.points(
    tri.data = ssc, geo = TT, bg='royalblue', 
    pch = 22, cex = 1, lwd = 1, 
    tri.sum.tst = FALSE
  )
  
  # simple legend
  legend('top', 
         legend = c('Source', 'Simulated'), 
         pch = c(22, 3), 
         col = c('black', 'firebrick'), 
         pt.bg = c('royalblue', NA), 
         horiz = TRUE, bty = 'n'
  )
  
  
}

}

}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.

Aitchison, J, C. Barcel'o-Vidal, J.J. Egozcue, V. Pawlowsky-Glahn (2002) A consise guide to the algebraic geometric structure of the simplex, the sample space for compositional data analysis, Terra Nostra, Schriften der Alfred Wegener-Stiftung, 03/2003
}
\author{
D.E. Beaudette
}
