% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilColorIndices.R
\name{buntley.westin.index}
\alias{buntley.westin.index}
\title{Buntley-Westin (1965) Index}
\usage{
buntley.westin.index(hue, chroma)
}
\arguments{
\item{hue}{A character vector containing Munsell hues (e.g. "7.5YR")}

\item{chroma}{A numeric vector containing Munsell chromas}
}
\value{
A numeric vector reflecting horizon redness (higher values = redder).
}
\description{
Calculate "Color Development Equivalent" by the method of Buntley & Westin (1965) "A Comparative Study of Developmental Color in a Chestnut-Chernozem-Brunizem Soil Climosequence" DOI: 10.2136/sssaj1965.03615995002900050029x. Originally developed for Mollisols, the Buntley-Westin index has been used as a tool to separate soils based on depth to particular colors.
}
\author{
Andrew G. Brown.
}
