% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glom.R
\name{clod.hz.ids}
\alias{clod.hz.ids}
\title{Return index to all horizons occuring in a depth interval}
\usage{
clod.hz.ids(p, z1, z2 = NA, modality = "all", as.list = FALSE)
}
\arguments{
\item{p}{A single-profile SoilProfileCollection}

\item{z1}{Top depth (required) - depth to intersect horizon at; if 'z2' specified, top depth of intersect interval.}

\item{z2}{Top depth (required) - depth to intersect horizon at; if 'z2' specified, top depth of intersect interval.}

\item{modality}{Return all data (default: "all") or first, thickest (\code{modality = "thickest"}) horizon in interval.}

\item{as.list}{OPTIONAL: return a list? default: FALSE}
}
\value{
A vector of unique horizon IDs.
}
\description{
\code{clod.hz.ids} returns a vector of unique indices corresponding to a depth interval. As arguments, it takes a single-profile SoilProfileCollection 'p', a top depth 'z1' and an optional bottom depth 'z2'.
 
If just top depth is specified, the unique index of the horizon intersected by that depth is returned. If bottom depth is specified, all horizon IDs that are intersected by the depth interval are returned.
}
\details{
This function is a workhorse used by glom() to create an index of unique horizon IDs intersecting the attribute of interest (depth) at the specified levels. This function returns the horizon IDs required to index an SPC and produce either a subset SPC or data.frame representation of the "clod" returned.
}
\examples{
data(sp1, package = 'aqp')
depths(sp1) <- id ~ top + bottom
site(sp1) <- ~ group
p <- sp1[1]
foo <- clod.hz.ids(p, 25, 100)
foo
}
\author{
Andrew G. Brown
}
