% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-coercion.R
\docType{methods}
\name{as}
\alias{as}
\alias{as,SoilProfileCollection-method}
\title{Coerce SoilProfileCollection with \code{as()}}
\value{
list

data.frame

tbl_df

data.table

SpatialPointsDataFrame

SpatialPoints
}
\description{
SoilProfileCollections can be coerced to other R object types using \code{as(spc, 'type')}.

Possible endpoints include: \code{list}, \code{data.frame}, \code{tbl_df}, \code{data.table}, \code{SpatialPointsDataFrame} and \code{SpatialPoints}.
}
\examples{
# load example data stored as SoilProfileCollection
data(sp5)

# list output
as(sp5, 'list')

# data.frame output
as(sp5, 'data.frame')

# tbl_df output
as(sp5, 'tbl_df')

# data.table output
as(sp5, 'data.table')

# SPDF output
as(sp5, 'SpatialPointsDataFrame')

# SP output
as(sp5, 'SpatialPoints')

}
