% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseHz.R
\name{collapseHz}
\alias{collapseHz}
\title{Collapse Horizons within Profiles Based on Pattern Matching}
\usage{
collapseHz(
  x,
  pattern,
  hzdesgn = guessHzDesgnName(x, required = TRUE),
  ignore.case = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{x}{A SoilProfileCollection}

\item{pattern}{character. A regular expression pattern to match in \code{hzdesgn} column}

\item{hzdesgn}{character. Any character column containing horizon-level identifiers. Default is estimated using \code{guessHzDesgnName()}.}

\item{ignore.case}{logical. If \code{FALSE}, the pattern matching is case sensitive and if \code{TRUE}, case is ignored during matching. Default: \code{FALSE}}

\item{na.rm}{logical. If \code{TRUE} \code{NA} values are ignored when calculating min/max boundaries for each group and in weighted averages. If \code{FALSE} \code{NA} values are propagated to the result. Default: \code{FALSE}}
}
\description{
Combines and aggregates layers by grouping adjacent horizons that match \code{pattern} in \code{hzdesgn}. Numeric properties are combined using the weighted average, and other properties are derived from the thickest horizon in each group.
}
\examples{
data(jacobs2000)

a <- collapseHz(jacobs2000, c(`A` = "^A", `E` = "E", `Bt` = "[ABC]+t", `C` = "^C", `foo` = "bar"))
b <- jacobs2000
profile_id(a) <- paste0(profile_id(a), "_collapse")

plot(c(a, b), color = "clay")
}
