% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\name{:=}
\alias{:=}
\title{:= (Assignment by reference MASK)}
\usage{
`:=`(...)
}
\arguments{
\item{...}{}
}
\description{
Mask to allow for Suggest of data.table \code{`:=`} / no R CMD CHECK error. See the manual page for \code{data.table::set} if interested in how this novel syntax is used.
}
\seealso{
\code{\link[data.table]{set}}
}
\keyword{internal}
