% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine,SoilProfileCollection-method}
\alias{combine,SoilProfileCollection-method}
\alias{combine}
\alias{combine,list-method}
\title{Combine SoilProfileCollection objects}
\usage{
\S4method{combine}{SoilProfileCollection}(...)

\S4method{combine}{list}(...)
}
\arguments{
\item{...}{SoilProfileCollection objects or lists of SoilProfileCollection objects to combine}
}
\value{
A SoilProfileCollection
}
\description{
Combine SoilProfileCollection objects or lists of SoilProfileCollection objects. This method provides \code{...} expansion for the \code{pbindlist} method.
}
\examples{

spc1 <- random_profile(1, SPC = TRUE)
spc2 <- random_profile(2, SPC = TRUE)

spc <- combine(spc1, spc2)

}
