% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeSPC.R
\name{compositeSPC}
\alias{compositeSPC}
\alias{compositeSPC,SoilProfileCollection-method}
\title{Return a list representation of site and horizon level data}
\usage{
compositeSPC(object)
}
\arguments{
\item{object}{A SoilProfileCollection}
}
\value{
A list.
}
\description{
\code{compositeSPC()} is a convenience function that returns a named list representation of the columns from the \code{@site} and \code{@horizons} slots.
}
\author{
Andrew G. Brown.
}
