% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{coordinates}
\alias{coordinates}
\alias{coordinates,SoilProfileCollection-method}
\title{Get coordinates from spatial slot}
\usage{
\S4method{coordinates}{SoilProfileCollection}(obj)
}
\arguments{
\item{obj}{a SoilProfileCollection}
}
\description{
Get coordinates from spatial slot, if present.
}
