% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{depth_units}
\alias{depth_units}
\alias{depth_units,SoilProfileCollection-method}
\alias{depth_units<-}
\alias{depth_units<-,SoilProfileCollection-method}
\title{Get depth units from metadata}
\arguments{
\item{object}{A SoilProfileCollection}

\item{value}{character, a value representing units. Default \code{'cm'}.}
}
\description{
Get units of depth measurement from metadata. Default value is centimeters.
}
\examples{

## set alternate depth units
du <- depth_units(sp1)

# inches
depth_units(sp1) <- 'in'

# replace original value (cm)
depth_units(sp1) <- du

}
