% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{dissolve_hz}
\alias{dissolve_hz}
\title{Dissolving horizon boundaries by grouping variables}
\usage{
dissolve_hz(
  object,
  by,
  collapse = FALSE,
  id = "peiid",
  top = "hzdept",
  bot = "hzdepb",
  order = FALSE
)
}
\arguments{
\item{object}{a \code{data.frame}}

\item{by}{character: column names, to be used as grouping variables, within the object.}

\item{collapse}{logical: indicating whether to not combine grouping variables before dissolving.}

\item{id}{character: column name of the pedon ID within the object.}

\item{top}{character: column name of the horizon top depth within the object.}

\item{bot}{character: column name of the horizon bottom depth in the object.}

\item{order}{logical: indicating whether or not to order the object by the id, top, and bot columns.
#'}
}
\value{
A \code{data.frame} with the original id, by grouping variables, and non-consecutive horizon depths.
}
\description{
This function dissolves or combines horizons share have a common set of grouping variables. It only combines those horizon records that are sequential (e.g. share a horizon boundary). Thus, it can be used to identify discontinuities in the grouping variables along a profile.
}
\details{
This function assumes the profiles and horizons within the object follow the logic defined by \code{checkHzDepthLogic} (e.g. records are ordered sequentially by id, top, and bot and without gaps). If the records are not order, set the argument \code{order = TRUE}.
}
\examples{

# example data
data(jacobs2000)

spc <- jacobs2000

spc$dep_5 <- spc$depletion_pct >=5
spc$genhz <- generalize.hz(spc$name, c("A", "E", "B", "C"), c("A", "E", "B", "C")) 
h <- horizons(spc)

test <- dissolve_hz(h, by = c("genhz", "dep_5"), id = "id", top = "top", bot = "bottom")

head(h[c("id", "top", "bottom", "genhz", "dep_5")])
head(test)

}
\seealso{
\code{\link{checkHzDepthLogic}}
}
\author{
Stephen Roecker
}
