% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-operators.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,SoilProfileCollection-method}
\title{Set data in column of horizon or site data in a SoilProfileCollection}
\usage{
\S4method{$}{SoilProfileCollection}(x, name) <- value
}
\arguments{
\item{x}{a SoilProfileCollection}

\item{name}{a single column name in site or horizon table}

\item{value}{Replacement values: unit length or equal to number of horizons or sites.}
}
\description{
Set the data in a column accessed by name \code{spc$name}. Column names other than profile ID are not shared between site and horizons. 

When using \code{$<-}, the length of input and output matching either the number of sites or number of horizons is used to determine which slot new columns are assigned to. Use \code{`site(x)$name <- value`} or  \code{`horizons(x)$name <- value`} to be explicit about which slot is being accessed.
}
