% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-operators.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,SoilProfileCollection-method}
\title{Get data from column of horizon or site data in a SoilProfileCollection}
\usage{
\S4method{$}{SoilProfileCollection}(x, name)
}
\arguments{
\item{x}{a SoilProfileCollection}

\item{name}{a single column name in site or horizon table}
}
\description{
Get the data from a column accessed by name \code{x$name}. Column
names other than profile ID are not shared between site and horizons.
Accessing the profile ID column (\code{idname(x)}) with \code{$} returns the
horizon-level data with profile ID repeated for each horizon in a profile.
}
\examples{

data(sp1)

depths(sp1) <- id ~ top + bottom

# get data from a column by name (prop)
sp1$prop

}
\seealso{
\code{\link[=site]{site()}} \code{\link[=horizons]{horizons()}} \code{\link[=idname]{idname()}} \code{\link[=profile_id]{profile_id()}}
}
