% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\name{.as.data.frame.aqp}
\alias{.as.data.frame.aqp}
\title{Wrapper method for data.frame subclass conversion}
\usage{
.as.data.frame.aqp(x, as.class = "data.frame", ...)
}
\arguments{
\item{x}{ANY.}

\item{as.class}{\code{"data.frame"}, \code{"tibble"}, or \code{"data.table"} default: \code{"data.frame"}}

\item{...}{Additional arguments to coercion function  \code{as.data.frame}, \code{as_tibble} or \code{as.data.table}}
}
\value{
a subclass of \code{data.frame} corresponding to \code{as.class},
}
\description{
Wrapper method for data.frame subclass conversion
}
