% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilColorSignature.R
\name{.colorSig2PerceptualDistMat}
\alias{.colorSig2PerceptualDistMat}
\title{Interpret a color signature containing color groups of CIELAB coordinates using perceptual distance via CIE dE00}
\usage{
.colorSig2PerceptualDistMat(pig)
}
\arguments{
\item{pig}{\code{data.frame} results from \verb{soilColorSignature(..., method = c('pam', 'depthSlices))}}
}
\value{
dist object
}
\description{
Interpret a color signature containing color groups of CIELAB coordinates using perceptual distance via CIE dE00
}
