% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2Munsell.R
\name{.detectColorSpec}
\alias{.detectColorSpec}
\title{Detect color specification from a vector of values, or a matrix of color coordinates}
\usage{
.detectColorSpec(col)
}
\arguments{
\item{col}{character vector, numeric matrix with 3 columns, or data.frame with 3 columns}
}
\description{
Detect color specification from a vector of values, or a matrix of color coordinates
}
