% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicedHSD.R
\name{.HSD}
\alias{.HSD}
\title{.HSD}
\usage{
.HSD(z, aov.fm, conf = 0.95)
}
\arguments{
\item{z}{data.frame containing basic metadata, horizon top/bottom, variable of interest, and grouping variable}

\item{aov.fm}{formula suitable for aov(): variable ~ group}

\item{conf}{confidence level for TukeyHSD}
}
\description{
Safely compute Tukey's HSD within a given slice
}
