% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarMunsellChips.R
\name{.makeEquivalentMunsellLUT}
\alias{.makeEquivalentMunsellLUT}
\title{.makeEquivalentMunsellLUT}
\usage{
.makeEquivalentMunsellLUT(threshold = 0.001)
}
\arguments{
\item{threshold}{Quantile cutoff for "equivalent" colors, default \code{0.001}}
}
\value{
A list with equal length to the number of rows (chips) in \code{munsell} data, each containing a numeric vector of row indices that are equivalent (CIE2000 distance less than \code{threshold}).
}
\description{
Makes the look up table based on pair-wise CIE2000 comparison of LAB colors for all whole value/chroma "chips" in the \code{munsell} data set via \code{farver::compare_colour}. Specify a threshold in terms of a probability level of CIE2000 distance (relative to whole dataset).
}
\examples{
\dontrun{

  # use default threshold
  equivalent_chip_lut <- .makeEquivalentMunsellLUT()

  # inspect the first 10 chips, it seems to work!
  lapply(equivalent_chip_lut[1:10], function(i) munsell[i,])

  # lets see some info on the number of chips per chip
  nchipsper <- sapply(equivalent_chip_lut, length)

  # top 10 are very high chroma chips with over 70 chips "identical"
  nchipsper[order(nchipsper, decreasing = TRUE)[1:10]]

  # look at distribution
  plot(density(nchipsper))

  # median is 5 -- Q: is this true of the range of Munsell colors typically used for soils?
  quantile(nchipsper)

  # double the default threshold
  doubletest <- sapply(.makeEquivalentMunsellLUT(threshold = 0.002), length)
  lines(density(doubletest), lty=2)

  # apprx. doubles the number of chips per chip in IQR
  quantile(doubletest)
}

}
\references{
Thomas Lin Pedersen, Berendea Nicolae and Romain François (2020). farver: High Performance Colour Space Manipulation. R package version 2.0.3. https://CRAN.R-project.org/package=farver
}
