% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicedHSD.R
\name{.parseGrouped_formula}
\alias{.parseGrouped_formula}
\title{.parseHSD_formula}
\usage{
.parseGrouped_formula(fm)
}
\arguments{
\item{fm}{to parse, 0:100 ~ variable | group}
}
\value{
list with formula pieces
}
\description{
internally used function to parse slicedHSD formula notation
}
