% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarMunsellChips.R
\name{equivalentMunsellChips}
\alias{equivalentMunsellChips}
\title{Identify "equivalent" (whole number value/chroma) Munsell chips
'}
\usage{
equivalentMunsellChips(hue = NULL, value = NULL, chroma = NULL)
}
\arguments{
\item{hue}{A character vector containing Munsell hues}

\item{value}{A numeric vector containing Munsell values (integer only)}

\item{chroma}{A numeric vector containing Munsell chromas (integer only)}
}
\value{
A named list; Each list element contains a data.frame with one or more rows of "equivalent" Munsell, RGB and LAB color coordinates from \code{munsell} data set.
}
\description{
Uses a pre-calculated look up table (via \code{farver::compare_colour}) based on pair-wise CIE2000 comparison of LAB colors for all whole value/chroma "chips" in the \code{munsell} data set. The "equivalent" chips stored in the look up table are based (fairly arbitrarily) on the 0.1% percentile of CIE2000 distances within the 8467x8467 distance matrix.
}
\examples{

# 7.5YR 4/4 (the one and only)
equivalentMunsellChips("7.5YR", 4, 4)

# 7.5YR 1/1
equivalentMunsellChips("7.5YR", 1, 1)

# 10YR 6/8
equivalentMunsellChips("10YR", 6, 8)

# compare visually
foo <- equivalentMunsellChips("10YR", 6, 8)[[1]]
pie(1:nrow(foo), col=munsell2rgb(foo$hue, foo$value, foo$chroma))

}
\references{
Thomas Lin Pedersen, Berendea Nicolae and Romain François (2020). farver: High Performance Colour Space Manipulation. R package version 2.0.3. https://CRAN.R-project.org/package=farver
}
