% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSoilColor.R
\name{estimateSoilColor}
\alias{estimateSoilColor}
\title{Estimate dry soil colors from moist soil colors and vice versa.}
\usage{
estimateSoilColor(hue, value, chroma, sourceMoistureState = c("dry", "moist"))
}
\arguments{
\item{hue}{vector of Munsell hue ('10YR', '2.5Y', etc.)}

\item{value}{vector of Munsell value (2,2.5 2.5, 3, 5, 6, etc.)}

\item{chroma}{vector of Munsell chroma (2, 3, 4, etc.)}

\item{sourceMoistureState}{character, source colors are either 'dry' or 'moist'}
}
\value{
\code{data.frame} of estimated Munsell colors
}
\description{
Estimate dry soil colors from moist soil colors and vice versa.
}
\details{
Pending
}
\examples{

estimateSoilColor(hue = '10YR', value = 3, chroma = 3, sourceMoistureState = 'moist')

}
\author{
D.E. Beaudette
}
