% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillHzGaps.R
\name{fillHzGaps}
\alias{fillHzGaps}
\title{Find and Fill Horizon Gaps}
\usage{
fillHzGaps(x, flag = FALSE, to_top = 0, to_bottom = max(x))
}
\arguments{
\item{x}{\code{SoilProfileCollection} object}

\item{flag}{logical, flag empty horizons that have been added. default: \code{FALSE}}

\item{to_top}{numeric, fill from shallowest top depth in each profile to specified depth? default: \code{0}}

\item{to_bottom}{numeric, fill from deepest bottom depth in each profile to specified depth? default: \code{aqp::max(x)}}
}
\value{
\code{SoilProfileCollection} object
}
\description{
This function attempt to find "gaps" in the horizon records of a \code{SoilProfileCollection} object and fill with place holder horizons.

Gaps are defined as:
\itemize{
\item within each profile, for horizons \code{i} to \code{n_hz}:
\item \verb{bottom_i != top_i+1 (but only to i = 1:(n_hz - 1)}
}
}
\examples{

data(sp4)
depths(sp4) <- id ~ top + bottom

# introduce depth logic errors
idx <- c(2, 6:7, 8, 12)
sp4$top[idx] <- NA

# check
horizons(sp4)[idx, ]

# create gaps by removing logic errors
x <- HzDepthLogicSubset(sp4, byhz = TRUE)

# inspect
par(mar = c(0, 0, 0, 1))
plotSPC(x, width = 0.3, default.color = 'royalblue', name = 'hzID')

z <- fillHzGaps(x, flag = TRUE)

# BUG: plotSPC can't use logical data for color
z$.filledGap <- as.factor(z$.filledGap)
plotSPC(z, width = 0.3, color = '.filledGap', name = 'hzID', show.legend = FALSE)

# fill top to 0 cm
z2 <- fillHzGaps(x, flag = TRUE, to_top = 0)
z2$.filledGap <- as.factor(z2$.filledGap)
plotSPC(z2, width = 0.3, color = '.filledGap', name = 'hzID', show.legend = FALSE)

# fill bottom to max(SPC)
z3 <- fillHzGaps(x, flag = TRUE, to_top = 0, to_bottom = max(x))
z3$.filledGap <- as.factor(z3$.filledGap)
plotSPC(z3, width = 0.3, color = '.filledGap', name = 'hzID', show.legend = FALSE)

}
