% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{filter}
\alias{filter}
\alias{filter,SoilProfileCollection-method}
\title{Subset a SoilProfileCollection with logical expressions}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{Comma-separated set of R expressions that evaluate as TRUE or FALSE. Length for individual expressions matche number of sites OR number of horizons, in \code{object}.}

\item{greedy}{Use "greedy" matching for combination of site and horizon level matches? \code{greedy=TRUE} is the union, whereas \code{greedy=FALSE} (default) is intersection}
}
\value{
A SoilProfileCollection.
}
\description{
\code{filter()} is a function used for subsetting SoilProfileCollections. It allows the user to specify an arbitrary number of logical vectors (equal in length to site or horizon), separated by commas. The function includes some support for "tidy" lexical features -- specifically, access to site and horizon-level variables directly by name.
}
\author{
Andrew G. Brown.
}
