% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap-fixOverlap.R
\name{findOverlap}
\alias{findOverlap}
\title{Find Overlap within a Sequence}
\usage{
findOverlap(x, thresh)
}
\arguments{
\item{x}{vector of relative horizontal positions, one for each profile}

\item{thresh}{threshold defining "overlap", typically < 1, ideal values likely in (0.3, 0.8)}
}
\value{
Unique indices of affected (overlapping) elements in \code{x}
}
\description{
Establish which elements within a vector of horizontal positions overlap beyond a given threshold
}
\examples{

x <- c(1, 2, 3, 3.4, 3.5, 5, 6, 10)

findOverlap(x, thresh = 0.5)

}
