% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap-fixOverlap.R
\name{fixOverlap}
\alias{fixOverlap}
\title{Fix overlap via random perturbation of affected elements}
\usage{
fixOverlap(
  x,
  thresh = 0.6,
  adj = 0.2,
  min.x = 0.8,
  max.x = length(x) + 0.2,
  maxIter = 1000,
  trace = FALSE
)
}
\arguments{
\item{x}{vector of horizontal positions}

\item{thresh}{horizontal threshold defining "overlap", must be < 1, ideal values likely in (0.3, 0.8)}

\item{adj}{adjustments are tested within \code{runif(min = adj * -1, max = adj)}}

\item{min.x}{left-side boundary condition}

\item{max.x}{right-side boundary condition}

\item{maxIter}{maximum number of iterations to attempt before giving up and returning integer sequence}

\item{trace}{print diagnostics}
}
\value{
\itemize{
\item \code{findOverlap} a vector of the same length as \code{x}, preserving rank-ordering and boundary conditions.
}
}
\description{
This function is actually quite stupid as it can converge on bogus results scaled adjustments based on deviation from threshold distances would be better or, SANN style adjustments
}
\examples{

x <- c(1, 2, 3, 3.4, 3.5, 5, 6, 10)

fixOverlap(x, thresh = 0.2, max.x = 10, trace = TRUE)

}
\author{
D.E. Beaudette
}
