% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastHorizonID.R
\name{getLastHorizonID}
\alias{getLastHorizonID}
\title{Get IDs of Deepest Horizons by Profile}
\usage{
getLastHorizonID(x)
}
\arguments{
\item{x}{a \code{SoilProfileCollection}}
}
\description{
Return horizon IDs of the deepest horizon within each profile of a \code{SoilProfileCollection}. IDs are returned in the same order as \code{profile_id(x)}. Horizon top depths are used because there are cases where bottom depths may be missing.
}
