% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glomApply.R
\name{glomApply}
\alias{glomApply}
\alias{glomApply,SoilProfileCollection-method}
\title{Subset an SPC by applying glom to each profile}
\usage{
glomApply(object, .fun = NULL, ..., chunk.size = 100)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{A set of comma-delimited R expressions that resolve to a transformation to be applied to a single profile e.g \code{glomApply(hzdept = max(hzdept) - hzdept)}}
}
\value{
A SoilProfileCollection.
}
\description{
\code{glomApply()} is a function used for subsetting SoilProfileCollections. A function \code{.fun} is supplied which is expected to return a top and bottom depth when supplied with a single-profile SPC and ... (any additional parameters). This numeric vector of length two is used by \code{glom()} to produce a subset SPC that is just the horizons overlapping the specified interval.
}
\author{
Andrew G. Brown.
}
