% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisisnotapipe.R
\name{grepSPC}
\alias{grepSPC}
\title{Subset SPC with pattern-matching for text-based attributes}
\usage{
grepSPC(object, attr, pattern, ...)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{attr}{A character vector (column in object) for matching patterns against.}

\item{pattern}{REGEX pattern to match in \code{attr}}

\item{...}{Additional arguments are passed to \code{grep()}}
}
\value{
A SoilProfileCollection.
}
\description{
\code{grepSPC()} is a function used for substting SoilProfileCollections in magrittr "pipelines." It includes some support for "tidy" lexical features -- namely access to the columns of site and horizon slots directly by name.
}
\author{
Andrew G. Brown.
}
