% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\alias{group_by,SoilProfileCollection-method}
\title{[EXPERIMENTAL] Store groupings within a profile collection.}
\usage{
\S4method{group_by}{SoilProfileCollection}(object, ...)
}
\arguments{
\item{object}{SoilProfileCollection.}

\item{...}{One or more expressions evaluated within the context of \code{object} that resolve to vectors that can be coerced to factor "groups."}
}
\description{
[EXPERIMENTAL] Store groupings within a profile collection.
}
