% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{horizonDepths}
\alias{horizonDepths}
\alias{horizonDepths,SoilProfileCollection-method}
\alias{horizonDepths<-}
\alias{horizonDepths<-,SoilProfileCollection-method}
\title{Get horizon depth column names}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a character vector of length two with names of columns containing numeric top and bottom depths}
}
\description{
Get column names containing horizon depths

Set column name containing horizon ID
}
