% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{horizonNames<-}
\alias{horizonNames<-}
\alias{horizonNames<-,SoilProfileCollection-method}
\alias{horizonNames}
\alias{horizonNames,SoilProfileCollection-method}
\title{Set horizon column names}
\arguments{
\item{value}{a unique vector of equal length to number of columns in horizons \code{length(horizonNames(object))}}

\item{object}{a SoilProfileCollection}
}
\description{
Set horizon column names

Get names of columns in horizon table.
}
