% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilColorIndices.R
\name{hurst.redness}
\alias{hurst.redness}
\title{Hurst (1977) Redness Index}
\usage{
hurst.redness(hue, value, chroma)
}
\arguments{
\item{hue}{A character vector containing Munsell hues (e.g. "7.5YR")}

\item{value}{A numeric vector containing Munsell values}

\item{chroma}{A numeric vector containing Munsell chromas}
}
\value{
A numeric vector of horizon redness index (lower values = redder).
}
\description{
Calculate Redness Index after Hurst (1977) "Visual estimation of iron in saprolite" DOI: 10.1130/0016-7606(1977)88<174:VEOIIS>2.0.CO;2. Accepts vectorized inputs for hue, value and chroma, produces vector output.
}
\references{
Hurst, V.J. (1977) Visual estimation of iron in saprolite. GSA Bulletin. 88(2): 174–176. doi: https://doi.org/10.1130/0016-7606(1977)88<174:VEOIIS>2.0.CO;2
}
\author{
Andrew G. Brown
}
