% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{hzDesgn}
\alias{hzDesgn}
\alias{hzDesgn,SoilProfileCollection-method}
\title{Get horizon designation column name}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get horizon designation names
}
