% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzdesgnname,SoilProfileCollection-method}
\alias{hzdesgnname,SoilProfileCollection-method}
\alias{hzdesgnname}
\title{Get horizon designation column name}
\usage{
\S4method{hzdesgnname}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get column name containing horizon designation name
}
