% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzID<-}
\alias{hzID<-}
\alias{hzID<-,SoilProfileCollection-method}
\title{Set horizon IDs}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of horizons \code{nrow(object)}}
}
\description{
Set vector containing horizon IDs
}
