% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzID<-,SoilProfileCollection-method}
\alias{hzID<-,SoilProfileCollection-method}
\alias{hzID<-}
\alias{hzID,SoilProfileCollection-method}
\alias{hzID}
\title{Set horizon IDs}
\usage{
\S4method{hzID}{SoilProfileCollection}(object) <- value

\S4method{hzID}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of horizons \code{nrow(object)}}
}
\description{
Set vector containing horizon IDs

Get vector containing horizon IDs
}
