% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{hzidname}
\alias{hzidname}
\alias{hzidname,SoilProfileCollection-method}
\alias{hzidname<-}
\alias{hzidname<-,SoilProfileCollection-method}
\title{Get horizon ID column name}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{character, column name containing unique horizon ID values}
}
\description{
Get column name containing unique horizon ID

Set unique horizon ID column name
}
\examples{
data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# create new horizon ID
sp1$hzIDrev <- rev(sp1$hzID)

# set horizon designation column
hzidname(sp1) <- "hzIDrev"

# get horizon designation column
hzidname(sp1)

}
