% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzMetadata,SoilProfileCollection-method}
\alias{hzMetadata,SoilProfileCollection-method}
\alias{hzMetadata}
\title{Get horizon-level metadata}
\usage{
\S4method{hzMetadata}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get \code{idname(object)} and \code{hzidname(object)}, with \code{hzdesgnname(object)}, \code{hztexclname(object)} (if defined)
}
