% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hztexclname<-}
\alias{hztexclname<-}
\alias{hztexclname<-,SoilProfileCollection-method}
\title{Set horizon texture class column name}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{character, name of column containing horizon texture classes}
}
\description{
Set horizon texture class column name for a SoilProfileCollection
}
\examples{

data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# set horizon texture class column
hztexclname(sp1) <- "texture"

# get horizon texture class column
hztexclname(sp1)
}
