% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{hztexclname}
\alias{hztexclname}
\alias{hztexclname,SoilProfileCollection-method}
\alias{hztexclame<-}
\title{Get horizon texture class column name}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{character, name of column containing horizon texture classes}
}
\description{
Get column name containing horizon designation name

Set horizon texture class column name
}
\examples{

data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# set horizon texture class column
hztexclname(sp1) <- "texture"

# get horizon texture class column
hztexclname(sp1)
}
