% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzTransitionSpline.R
\name{hzTransitionSplineSolve}
\alias{hzTransitionSplineSolve}
\title{Mass-preserving Spline Solver}
\usage{
hzTransitionSplineSolve(L, dt, db, vals, k, vlow, vhigh)
}
\arguments{
\item{L}{Horizon thicknesses (cm)}

\item{dt}{Top transition half-widths (cm)}

\item{db}{Bottom transition half-widths (cm)}

\item{vals}{Horizon values}

\item{k}{factor (1/3 for quadratic, 1/4 for linear)}

\item{vlow}{Minimum allowable value}

\item{vhigh}{Maximum allowable value}
}
\value{
A list containing \code{V} (constant values for each horizon) and \code{Y_all} (values at each horizon boundary)
}
\description{
Internal solver for mass-preserving spline with transition zones.
}
