% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{idname}
\alias{idname}
\alias{idname,SoilProfileCollection-method}
\title{Get profile ID column name}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get column name containing unique profile IDs
}
