% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_profiles.R
\name{L1_profiles}
\alias{L1_profiles}
\title{Create Representative Profiles via L1 Median}
\usage{
L1_profiles(
  x,
  fm,
  basis = 1,
  method = c("regex", "simple", "constant"),
  maxDepthRule = c("max", "min"),
  maxDepthConstant = NULL,
  strict = FALSE
)
}
\arguments{
\item{x}{\code{SoilProfileCollection} object}

\item{fm}{formula}

\item{basis}{aggregation basis}

\item{method}{soil depth evaluation method: regular expression, simple, constant. See details.}

\item{maxDepthRule}{maximum depth rule: min | max. See details.}

\item{maxDepthConstant}{depth when \code{maxDepthRule = 'constant'}}

\item{strict}{passed to \code{slice}}
}
\value{

}
\description{
Create Representative Profiles via L1 Median
}
