% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{max}
\alias{max}
\alias{max,SoilProfileCollection-method}
\title{Get the maximum bottom depth in a SoilProfileCollection}
\usage{
\S4method{max}{SoilProfileCollection}(x, v = NULL)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get the deepest depth of description out of all profiles in a SoilProfileCollection.
}
