% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixMunsell.R
\name{mixMunsell}
\alias{mixMunsell}
\title{Mix Munsell Colors via Spectral Library}
\usage{
mixMunsell(x, w = c(0.5, 0.5))
}
\arguments{
\item{x}{vector of two colors in Munsell notation}

\item{w}{vector of two weights or proportions, can sum to any number}
}
\value{
a list containing:

\itemize{

\item{fig}{lattice graphic}
\item{mixed}{a \code{data.frame} with the closest matching Munsell color}

}
}
\description{
Simulate subtractive mixing (e.g. pigments) of two colors in Munsell notation.
}
\details{
Pending
}
\references{
inspiration / calculations based on:
https://arxiv.org/ftp/arxiv/papers/1710/1710.06364.pdf
related discussion here:
https://stackoverflow.com/questions/10254022/implementing-kubelka-munk-like-krita-to-mix-colours-color-like-paint/29967630#29967630
base spectral library:
http://www.munsellcolourscienceforpainters.com/MunsellResources/SpectralReflectancesOf2007MunsellBookOfColorGlossy.txt
}
\author{
D.E. Beaudette
}
