% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarMunsellChips.R
\docType{data}
\name{munequivalent}
\alias{munequivalent}
\title{Indices of "equivalent" Munsell chips in \code{munsell} data set}
\format{
A named list with 8467 elements, each containing a numeric vector of indices corresponding to the \code{munsell} data set, which has 8467 rows (unique, whole-number chips). Names have the format \code{HUE VALUE/CHROMA}, eg. \code{"7.5YR 4/4"}
}
\usage{
data(munequivalent)
}
\description{
A pre-calculated look up table (made with \code{farver::compare_colour}) based on pair-wise CIE2000 comparison of LAB colors for all whole value/chroma "chips" in the \code{munsell} data set.

"Equivalent" chips table are based (fairly arbitrarily) on the 0.1% percentile of CIE2000 distances within the 8467x8467 distance matrix.
}
\references{
Thomas Lin Pedersen, Berendea Nicolae and Romain François (2020). farver: High Performance Colour Space Manipulation. R package version 2.0.3. https://CRAN.R-project.org/package=farver
}
\keyword{datasets}
