% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{munsell.spectra}
\alias{munsell.spectra}
\alias{munsell.spectra.wide}
\title{Spectral Library of Munsell Colors}
\format{
A data frame with 89496 rows and 10 variables:
\describe{
  \item{munsell}{munsell color}
  \item{hue}{hue component}
  \item{value}{value component}
  \item{chroma}{chroma component}
  \item{wavelength}{wavelength (nm)}
  \item{reflectance}{reflectance}
}
}
\usage{
data(munsell.spectra)
}
\description{
The original database (\href{http://centore.isletech.net/MunsellResources/SpectralReflectancesOf2007MunsellBookOfColorGlossy.txt}{SpectralReflectancesOf2007MunsellBookOfColorGlossy.txt}) was provided by Paul Centore and downloaded September, 2020. Reflectance values for odd chroma have been interpolated from adjacent chips. See \code{aqp/misc/utils/Munsell/} for the entire set of processing steps.

The original database contains the following description:

This file contains spectral reflectance measurements of X-Rite's 2007 Munsell Book of Color (Glossy Finish).  The measurements were made in 2012 with a ColorMunki spectrophotometer.  The first column is the Munsell name.  The remaining columns give reflectance values for 380 nm to 730 nm, in steps of 10 nm.  The reflectance is a value between 0 (indicating that no light at that wavelength is reflected) and 1 (indicating that all the light at that wavelength is reflected).  Occasionally an entry is slightly greater than 1.  The likely cause is random variability, and those entries can be adjusted to 1 with negligible loss. In all, 1485 colour samples were measured.  Researchers are invited to analyze the data in this file.
}
\references{
Centore, Paul. Colour Tools for Painters. \url{http://www.munsellcolourscienceforpainters.com/}.
}
\keyword{datasets}
