% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{munsellHuePosition}
\alias{munsellHuePosition}
\title{Munsell Hue Position Reference}
\format{
An object of class \code{data.frame} with 41 rows and 4 columns.
}
\usage{
data(munsellHuePosition)
}
\description{
Position data for the 40 standard Munsell hues (and neutral). Data include angular positions (compass-style, origin at \verb{[x = 0, y = 1]}, CW rotation) and Cartesian coordinates on the unit circle.
}
\references{
Munsell book of color. 1976. Macbeth, a Division of Kollmorgen Corp., Baltimore, MD.
}
\keyword{datasets}
