% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{nrow,SoilProfileCollection-method}
\alias{nrow,SoilProfileCollection-method}
\title{Get the number of horizons in a SoilProfileCollection}
\usage{
\S4method{nrow}{SoilProfileCollection}(x)
}
\arguments{
\item{x}{a SoilProfileCollection}
}
\description{
Get the number of horizons in a SoilProfileCollection
}
