% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Ohz.colors}
\alias{Ohz.colors}
\title{Representative O Horizon Colors}
\format{
A \code{data.frame} with the following elements:
\itemize{
\item state: soil moisture state, "dry" or "moist"
\item genhz: horizon designation generalized into one of "Oi", "Oe", "Oa", or "other"
\item L1.munsell: representative color in Munsell notation for each moisture state / generalized horizon designation
\item group.dE00: weighted-mean CIE dE 2000 color contrast metric between all soil colors in this group and representative color
}
}
\usage{
data(Ohz.colors)
}
\description{
Representative O horizon colors for the most common types of organic soil material, at dry and moist states. Colors were derived from an analysis of soil morphologic data (5245 horizons) within the USDA-NRCS National Soil Information System. Representative colors are the L1 median colors (by \code{\link[=colorVariation]{colorVariation()}}) within groups generalized to "Oi", "Oe", "Oa", and all "other". These estimates may be useful place-holder values for soil color in collections where O horizon color was not recorded.
}
\keyword{datasets}
