% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{osd}
\alias{osd}
\title{Example Output from soilDB::fetchOSD()}
\format{
A \code{SoilProfileCollection}
}
\usage{
data(osd)
}
\description{
An example \code{SoilProfileCollection} object created by \code{soilDB::fetchOSD()}, derived from the Cecil, Appling, and Bonneau Official Series Descriptions.
}
\keyword{datasets}
