% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqp-label-placement-solvers.R
\name{overlapMetrics}
\alias{overlapMetrics}
\title{Find and Quantify Overlap within a 1D Sequence}
\usage{
overlapMetrics(x, thresh)
}
\arguments{
\item{x}{vector of relative horizontal positions, one for each profile}

\item{thresh}{threshold defining "overlap", typically < 1

@return a \code{list}:
\itemize{
\item \code{idx}: unique index to overlapping elements in \code{x}
\item \code{ov}: normalized overlap (see details)
}}
}
\description{
Desc.
}
\examples{

x <- c(1, 2, 3, 3.4, 3.5, 5, 6, 10)

overlapMetrics(x, thresh = 0.5)


}
