% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb2.R
\name{perturb2}
\alias{perturb2}
\title{perturb2}
\usage{
perturb2(x, n = 100, sds = rep(1, nrow(x)))
}
\arguments{
\item{x}{a SoilProfileCollection}

\item{n}{number of realizations}

\item{sds}{vector of standard deviations, or a column name in the horizon slot of x}
}
\value{
a SoilProfileCollection with \code{length(x) * n} profiles
}
\description{
It's vectorized
}
\examples{
data(sp5)
horizons(sp5)$sdvar <-  1

# old perturb
system.time(res1 <- combine(profileApply(sp5[1:5,], function(x) {
  res <- aqp::perturb(x, boundary.attr = "sdvar")
  profile_id(res) <- paste0(profile_id(x),'_',profile_id(res))
  res
})))

# new perturb
system.time(res2 <- perturb2(sp5[1:5,]))
}
