% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPC.R
\name{plot,SoilProfileCollection,ANY-method}
\alias{plot,SoilProfileCollection,ANY-method}
\title{generic plot method for \code{SoilProfileCollection} objects}
\usage{
\S4method{plot}{SoilProfileCollection,ANY}(
  x,
  color = "soil_color",
  width = 0.2,
  name = NULL,
  name.style = "right-center",
  label = idname(x),
  hz.depths = FALSE,
  alt.label = NULL,
  alt.label.col = "black",
  cex.names = 0.5,
  cex.depth.axis = cex.names,
  cex.id = cex.names + (0.2 * cex.names),
  font.id = 2,
  print.id = TRUE,
  id.style = "auto",
  plot.order = 1:length(x),
  relative.pos = 1:length(x),
  add = FALSE,
  scaling.factor = 1,
  y.offset = 0,
  x.idx.offset = 0,
  n = length(x),
  max.depth = ifelse(is.infinite(max(x)), 200, max(x)),
  n.depth.ticks = 5,
  shrink = FALSE,
  shrink.cutoff = 3,
  abbr = FALSE,
  abbr.cutoff = 5,
  divide.hz = TRUE,
  hz.distinctness.offset = NULL,
  hz.topography.lty = NULL,
  axis.line.offset = -2.5,
  plot.depth.axis = TRUE,
  density = NULL,
  col.label = color,
  col.palette = rev(brewer.pal(10, "Spectral")),
  col.palette.bias = 1,
  col.legend.cex = 1,
  n.legend = 8,
  lwd = 1,
  lty = 1,
  default.color = grey(0.95),
  ...
)
}
\description{
generic plot method for \code{SoilProfileCollection} objects
}
