% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorQuantiles.R
\name{plotColorQuantiles}
\alias{plotColorQuantiles}
\title{Visualize Color Quantiles}
\usage{
plotColorQuantiles(res, pt.cex = 7, title = "", mar = c(2, 1.5, 1, 1))
}
\arguments{
\item{res}{list returned by \code{colorQuantiles}}

\item{pt.cex}{scaling factor for color chips}

\item{title}{optional title printed above color quantile figure}

\item{mar}{customized margins, see details}
}
\description{
This function creates a visualization of the output from \code{colorQuantiles}.
}
\author{
D.E. Beaudette
}
