% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{profile_id}
\alias{profile_id}
\alias{profile_ID,SoilProfileCollection-method,}
\alias{profile_ID<-,SoilProfileCollection-method}
\alias{profile_id<-}
\alias{profile_id<-,SoilProfileCollection-method}
\title{Get/set unique (sorted) profile IDs}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of profiles \code{length(object)}}
}
\description{
Get or set a vector of profile IDs

Set vector containing profile IDs
}
