% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-spatial.R
\name{proj4string<-,SoilProfileCollection,ANY-method}
\alias{proj4string<-,SoilProfileCollection,ANY-method}
\title{Set PROJ4 string for the SoilProfileCollection}
\usage{
\S4method{proj4string}{SoilProfileCollection,ANY}(obj) <- value
}
\arguments{
\item{obj}{A SoilProfileCollection}

\item{value}{A proj4string}
}
\description{
Set PROJ4 string for the SoilProfileCollection
}
