% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-spatial.R
\name{proj4string,SoilProfileCollection-method}
\alias{proj4string,SoilProfileCollection-method}
\title{Set PROJ4 string for the SoilProfileCollection}
\usage{
\S4method{proj4string}{SoilProfileCollection}(obj)
}
\arguments{
\item{obj}{A SoilProfileCollection}
}
\description{
Set PROJ4 string for the SoilProfileCollection
}
