% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture.R
\name{PSCS_levels}
\alias{PSCS_levels}
\title{Ranking Systems for USDA Taxonomic Particle-Size and Substitute Classes of Mineral Soils}
\usage{
PSCS_levels()
}
\value{
an ordered factor
}
\description{
Generate a vector of USDA Particle-Size and Substitute Classes  names, sorted according to approximate particle size
}
\examples{

# class codes
PSCS_levels()

}
\references{
\href{https://nrcspad.sc.egov.usda.gov/DistributionCenter/product.aspx?ProductID=991}{Field Book for Describing and Sampling Soils, version 3.0}
}
\seealso{
\code{\link[=hz_to_taxpartsize]{hz_to_taxpartsize()}}, \code{\link[=texture_to_taxpartsize]{texture_to_taxpartsize()}}, \code{\link[=SoilTextureLevels]{SoilTextureLevels()}}
}
\author{
Stephen Roecker
}
