% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{rbind.SoilProfileCollection}
\alias{rbind.SoilProfileCollection}
\title{rbind.SoilProfileCollection}
\usage{
\method{rbind}{SoilProfileCollection}(...)
}
\arguments{
\item{...}{One or more SoilProfileCollection objects to \code{union()}}
}
\value{
A SoilProfileCollection
}
\description{
rbind.SoilProfileCollection
}
\examples{

data(sp5)

rbind(sp5[1:2,], sp5[(length(sp5) - 1):length(sp5),])

}
