% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{ph_to_rxnclass}
\alias{ph_to_rxnclass}
\alias{rxnclass_to_ph}
\title{Convert pH to/from Reaction Classes}
\usage{
ph_to_rxnclass(x, halfclass = FALSE)

rxnclass_to_ph(x, halfclass = FALSE, digits = 2, simplify = TRUE)
}
\arguments{
\item{x}{input pH values (numeric; \code{ph_to_rxnclass()}) or reaction classes (character; \code{rxnclass_to_ph()})}

\item{halfclass}{Split the standard classes in half for higher resolution? Default: \code{FALSE}}

\item{digits}{Number of digits after decimal place; Default: \code{2}. Used only for \code{rxnclass_to_ph()}}

\item{simplify}{Simplify list result to numeric vector when length of result is 1? Default: \code{TRUE}}
}
\value{
\code{ph_to_rxnclass()}: a vector of reaction classes

\code{rxnclass_to_ph()}: a list of high/low values of reaction class 1:1 with input; if simplify=TRUE and result is length 1, a numeric vector containing low and high values.
}
\description{
Convert pH to/from Reaction Classes
}
\examples{
ph_to_rxnclass(6.2)
rxnclass_to_ph("slightly acid")
}
