% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{reactionclass}
\alias{reactionclass}
\title{pH Reaction Classes}
\format{
An object of class \code{data.frame} with 11 rows and 3 columns.
}
\usage{
data(reactionclass)
}
\description{
Levels of pH (reaction) classes including descriptive name, and range from low to high pH
}
\references{
Soil Science Division Staff. (2017) Soil Survey Manual. C. Ditzler, K. Scheffe, and H.C. Monger (eds.). USDA Handbook 18. Government Printing Office, Washington, D.C.
}
\keyword{datasets}
