% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuildSPC.R
\name{rebuildSPC}
\alias{rebuildSPC}
\title{Rebuild a SoilProfileCollection object}
\usage{
rebuildSPC(x)
}
\arguments{
\item{x}{a \code{SoilProfileCollection} object}
}
\value{
A valid \code{SoilProfileCollection} object.

A valid \code{SoilProfileCollection} object.
}
\description{
Rebuild a SoilProfileCollection object
}
\details{
Attempt rebuilding a \code{SoilProfileCollection} object by splitting into
components and re-assembling. Likely only used to fix outdated
\code{SoilProfileCollection} objects that are missing slots.
}
\seealso{
\code{\link{checkSPC}}
Rebuild a SoilProfileCollection object

Rebuild a SoilProfileCollection object

Attempt rebuilding a \code{SoilProfileCollection} object by splitting into
components and re-assembling. Likely only used to fix outdated
\code{SoilProfileCollection} objects that are missing slots.

\code{\link{checkSPC}}
}
\author{
D.E. Beaudette

D.E. Beaudette, A.G. Brown
}
