% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-integrity.R
\docType{methods}
\name{reorderHorizons}
\alias{reorderHorizons}
\alias{reorderHorizons,SoilProfileCollection-method}
\title{Re-order corrupted horizon data}
\usage{
\S4method{reorderHorizons}{SoilProfileCollection}(object, target.order = NULL)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{target.order}{A numeric vector of equal length to \code{object}. Default value is \code{NULL} which restores the internal order of the collection.}
}
\value{
SoilProfileCollection
}
\description{
This is a method made available primarily to repair horizon data that have been corrupted relative to their order at time of SoilProfileCollection construction. 

There is an option to specify the target order, but this will not update the corresponding metadata entry tracking the original order. Use this functionality at your own risk.
}
