% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{rruff.sample}
\alias{rruff.sample}
\title{Sample XRD Patterns}
\format{
A data frame with 3000 observations on the following 8 variables.
\describe{ \item{twotheta}{twotheta values}
\item{nontronite}{XRD pattern for nontronite}
\item{montmorillonite}{XRD pattern for montmorillonite}
\item{clinochlore}{XRD pattern for clinochlore}
\item{antigorite}{XRD pattern for antigorite}
\item{chamosite}{XRD pattern for chamosite}
\item{hematite}{XRD pattern for hematite}
\item{goethite}{XRD pattern for goethite} }
}
\source{
http://rruff.info/
}
\usage{
data(rruff.sample)
}
\description{
Several sample XRD patterns from the RRUFF project site.
}
\examples{

data(rruff.sample)

# plot all patterns
matplot(rruff.sample, type='l', lty=1)


}
\references{
http://rruff.info/
}
\keyword{datasets}
