% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segmenting of Soil Horizon Data by Depth Interval}
\usage{
segment(object = NULL, intervals = NULL, trim = TRUE, hzdepcols = NULL)
}
\arguments{
\item{object}{either a \code{SoilProfileCollection} or \code{data.frame}}

\item{intervals}{a vector of integers over which to slice the horizon data (e.g. \code{c(25, 100)} or \code{25:100})}

\item{trim}{logical, specify whether horizon depths should be clipped to the depth intervals (e.g. 15 -> 25)}

\item{hzdepcols}{a character vector of length 2 specifying the names of the horizon depths (e.g. \code{c("hzdept", "hzdepb")}), only necessary if \code{object} is a \code{data.frame}.}
}
\value{
Either a \code{SoilProfileCollection} or \code{data.frame} with the original horizon data segmented by depth intervals. A new column called \code{segment_id} identifying the depth interval is added.
}
\description{
This function segments horizon data from both SoilProfileCollections and data frames into depth intervals or slices.
}
\details{
This function will segment soil horizon data by arbitrary depth intervals (e.g. c(25, 100) or 25:100). Compared to \code{slice}, segment is more generic, as it supports artibrary intervals, and works with either SoilProfileCollections or data.frames. In addition \code{segment} is slightly faster than \code{slice}. Compared to \code{glom}, \code{segment} works on more than a single SoilProfileCollection.

\code{segment} is intended to make it less computationally intensive to calculate weighted averages within specified depth intervals, because horizon thicknesses can be used as weights. Therefore it is not necessary to slice horizon data into 1-cm increments, which increases \code{object.size} a hundred fold or more, prior to computing averages or quantiles in order to properly weight the contribution from horizons of varying thicknesses.
}
\author{
Stephen Roecker
}
