% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{DEPRECATED Simulate Soil Profiles}
\usage{
sim(x, n = 1, iterations = 25, hz.sd = 2, min.thick = 2)
}
\arguments{
\item{x}{a SoilProfileCollection object containing a single profile from which to draw simulated data}

\item{n}{the number of requested simulations}

\item{iterations}{sampling iterations used to determine each horizon thickness}

\item{hz.sd}{standard deviation used to simulate horizon thickness, can be a vector but must divide evenly into the number of horizons found in \code{x}}

\item{min.thick}{minimum horizon thickness allowed in simulation results}
}
\value{
A SoilProfileCollection object with \code{n} simulated profiles, each containing the same number of horizons and same data as \code{x}
}
\description{
Simulate a collection of soil profiles based on the horizonation of a single soil profile. Now deprecated: use \code{\link[=perturb]{perturb()}} for perturbations of horizon thicknesses or boundaries.
}
\details{
This function generates a collection of simulated soil profiles based on the horizon thickness data associated with a single "template" profile. Simulation is based on sampling from a family of Gaussian distribution with means defined by the "template" profile and standard deviation defined by the user.
}
\examples{

# please see documentation for perturb() for examples
#  the sim() function calls perturb() internally

}
\seealso{
\code{\link{random_profile}}  \code{\link{perturb}}
}
\author{
D. E. Beaudette
}
