% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{siteNames}
\alias{siteNames}
\alias{siteNames,SoilProfileCollection-method}
\alias{siteNames<-}
\alias{siteNames<-,SoilProfileCollection-method}
\title{Get names of columns in site table}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of columns in site: \code{length(siteNames(object))}}
}
\description{
Get names of columns in site table.

Set site column names
}
