% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicedHSD.R
\name{slicedHSD}
\alias{slicedHSD}
\title{Tukey's HSD Over Slices}
\usage{
slicedHSD(object, fm, conf = 0.95)
}
\arguments{
\item{object}{\code{SoilProfileCollection} object}

\item{fm}{a formula describing depth sequence, horizon attribute, and site (grouping) attribute. 
For example 0:100 ~ estimated_oc | taxonname}

\item{conf}{confidence applied in \code{TukeyHSD}}
}
\description{
Apply Tukey's HSD over 1cm slices.
}
\author{
D.E. Beaudette and Sharon Perrone
}
