% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-spatial.R
\name{prj,SoilProfileCollection-method}
\alias{prj,SoilProfileCollection-method}
\alias{prj}
\alias{prj<-,SoilProfileCollection-method}
\alias{prj<-}
\alias{proj4string,SoilProfileCollection-method}
\alias{proj4string<-,SoilProfileCollection,ANY-method}
\title{Get or Set Coordinate Reference System for SoilProfileCollection}
\usage{
\S4method{prj}{SoilProfileCollection}(object, ...)

\S4method{prj}{SoilProfileCollection}(object, ...) <- value

\S4method{proj4string}{SoilProfileCollection}(obj)

\S4method{proj4string}{SoilProfileCollection,ANY}(obj) <- value
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{Additional arguments (not used)}

\item{value}{character. Representation of Coordinate Reference System as \code{"authority:code"}, integer EPSG code, WKT2019 / PROJ4 string, an sf \code{crs} or sp \code{CRS} object.}

\item{obj}{A SoilProfileCollection}
}
\description{
\code{prj()}: Get Coordinate Reference System (Projection) metadata

\verb{prj()<-}: Set Coordinate Reference System metadata for the SoilProfileCollection

\code{proj4string()}: (Deprecated) Get Coordinate Reference System as PROJ4 String

\verb{proj4string()<-}: (Deprecated) Set Coordinate Reference System metadata for the SoilProfileCollection
}
\seealso{
\code{\link[=initSpatial<-]{initSpatial<-()}}
}
