\name{spatial_subset-methods}
\docType{methods}
\alias{spatial_subset}
\alias{spatial_subset,SoilProfileCollection-method}
\title{Spatial Subsetting of SoilProfileCollection Objects}
\description{Spatial Subsetting of SoilProfileCollection Objects}
\section{Methods}{
\describe{

\item{\code{signature(object = "SoilProfileCollection")}}{}
}}
\note{This functionality requires the `rgeos` package.}
\keyword{methods}
\keyword{manip}
