% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{SPC.with.overlap}
\alias{SPC.with.overlap}
\title{Example SoilProfileCollection with Overlapping Horizons}
\format{
A \code{SoilProfileCollection}
}
\usage{
data(SPC.with.overlap)
}
\description{
A \code{SoilProfileCollection} with overlapping horizons, derived from a Dynamic Soil Properties project.
}
\keyword{datasets}
