% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_horizonOffset.R
\name{horizon_above}
\alias{horizon_above}
\alias{horizon_below}
\alias{spc_horizonOffset}
\title{Horizons Above/Below}
\usage{
horizon_above(x, ..., offset = 1, SPC = TRUE, simplify = SPC)

horizon_below(x, ..., offset = 1, SPC = TRUE, simplify = SPC)

spc_horizonOffset(x, hzid, offset, SPC = FALSE, simplify = TRUE)
}
\arguments{
\item{x}{A SoilProfileCollection}

\item{...}{Comma-separated set of R expressions that evaluate as \code{TRUE} or \code{FALSE} in context of horizon data frame. Length for individual expressions matches number of horizons, in \code{x}.}

\item{offset}{Integer offset in terms of SoilProfileCollection \verb{[,j]} (horizon/slice) index}

\item{SPC}{Return a SoilProfileCollection? Default \code{TRUE} for \verb{horizon_*} methods.}

\item{simplify}{If \code{TRUE} return a vector (all elements combined), or a list (1 element per profile). If \code{SPC} is \code{TRUE} then \code{simplify} is \code{TRUE}.}

\item{hzid}{A vector of target horizon IDs. These are calculated from \code{...} for \verb{horizon_*()} methods}
}
\value{
A SoilProfileCollection (when \code{SPC = TRUE}) or a vector of horizon row indices (when \code{SPC = FALSE} and \code{simplify = TRUE}) or a list (when \code{SPC = FALSE} and \code{simplify = FALSE}))
}
\description{
Horizons Above/Below
}
\examples{
data(sp4)
depths(sp4) <- id ~ top + bottom

# get the horizon above the last horizon (j-index of bottom horizon minus 1)
horizon_above(sp4, hzID(sp4) \%in\% getLastHorizonID(sp4))

# get horizons below the last horizon (none; j-index of bottom horizon plus 1)
horizon_below(sp4, hzID(sp4) \%in\% getLastHorizonID(sp4))

}
