% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2Munsell.R
\name{spec2Munsell}
\alias{spec2Munsell}
\title{Convert reflectance spectra to closest Munsell chip}
\usage{
spec2Munsell(x, ...)
}
\arguments{
\item{x}{reflectance spectra, (380nm to 730nm, 10nm resolution)}

\item{...}{further arguments to \code{rgb2Munsell}}
}
\value{
output from \code{rgb2Munsell}
}
\description{
Convert reflectance spectra to closest Munsell chip
}
\note{
D65 illuminant spectra and CIE1931 color matching functions subset from data provided by the {colorscience} package.
}
\references{
Wyszecki, G., & Stiles, W. S., 1982 Color Science: concepts and methods, quantitative data and formulae (2nd ed.). New York: Wiley.

Jose Gama and Glenn Davis (2019). colorscience: Color Science Methods and Data. R package version 1.0.8.
https://CRAN.R-project.org/package=colorscience
}
