% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-to-sharpshootR.R
\name{getArgillicBounds}
\alias{getArgillicBounds}
\alias{argillic.clay.increase.depth}
\alias{crit.clay.argillic}
\alias{estimatePSCS}
\alias{get.increase.depths}
\alias{get.increase.matrix}
\alias{getCambicBounds}
\alias{hasDarkColors}
\alias{mollic.thickness.requirement}
\title{Soil Taxonomy Heuristics (moved to sharpshootR package)}
\usage{
getArgillicBounds(...)
}
\arguments{
\item{...}{arguments to functions deprecated from aqp namespace}
}
\description{
Soil Taxonomy Heuristics (moved to sharpshootR package)
}
