% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisisnotapipe.R
\name{subSPC}
\alias{subSPC}
\title{Subset SPC with logical vectors}
\usage{
subSPC(object, ...)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{Comma-separated set of R expressions that evaluate as TRUE or FALSE. Length for individual expressions matche number of sites OR number of horizons, in \code{object}.}
}
\value{
A SoilProfileCollection.
}
\description{
\code{subSPC()} is a function used for substting SoilProfileCollections in magrittr "pipelines." It allows the user to specify an arbitrary number of logical vectors (equal in length to site or horizon), separated by commas. It includes some support for "tidy" lexical features -- specifically, access to site and horizon-level variables directly by name.
}
\author{
Andrew G. Brown.
}
