% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeSPC}
\alias{summarizeSPC}
\alias{summarizeSPC,SoilProfileCollection-method,}
\alias{summarize}
\title{Perform summaries on groups (from \code{group_by}) and create new site or horizon level attributes}
\usage{
summarizeSPC(object, ...)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{A set of (named) comma-delimited R expressions that resolve to a summary value. e.g \code{groupmean = mean(clay, na.rm = TRUE)}}
}
\value{
A data.frame with one row per level in the grouping variable, and one column for each summary
}
\description{
\code{summarize()} is a function used for summarizing SoilProfileCollections. Specify the groups using the group_by verb, and then (named) expressions to evaluate on each group. The result is a data.frame with one row per categorical level in the grouping variable and one column for each summary variable.
}
\author{
Andrew G. Brown
}
