% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{traditionalColorNames}
\alias{traditionalColorNames}
\title{Traditional Soil Color Names}
\format{
An object of class \code{data.frame} with 482 rows and 2 columns.
}
\usage{
data(traditionalColorNames)
}
\description{
Traditional soil color names associated with select Munsell colors.
}
\references{
Sourced from the "colorconverter" NASIS property script.
}
\keyword{datasets}
