% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_depths}
\alias{validate_depths}
\title{Check horizon depths for errors}
\usage{
validate_depths(
  object,
  id = "peiid",
  top = "hzdept",
  bot = "hzdepb",
  append_checks = TRUE,
  pad_bot = FALSE,
  rmHzErrors = FALSE
)
}
\arguments{
\item{object}{data.frame or SoilProfileCollection object to check}

\item{id}{pedon id column name, only necessary if object is a data.frame. Default: \code{peiid}}

\item{top}{horizon top depth column name, only necessary if object is a data.frame. Default: \code{hzdept}}

\item{bot}{horizon bottom depth column name, only necessary if object is a data.frame. Default: \code{hzdepb}}

\item{append_checks}{logical indicating whether the validation check should be appended to the input object. Default: \code{TRUE}}

\item{pad_bot}{logical indicating whether to pad the bottom horizon with the top + 1-cm if the top == bot or the bot is NA. Default: \code{FALSE}}

\item{rmHzErrors}{logical indicating whether to remove pedons with horizon errors from the output. Default: \code{FALSE}}
}
\value{
A \code{data.frame} or \code{SoilProfileCollection} appended with new \code{check_} columns appended if \code{append_check = TRUE}.
\itemize{
\item \code{valid_dep_order}: \code{TRUE} if \code{top < bot}
\item \code{valid_dep_diff}: \code{TRUE} if \code{top != bot}
\item \code{valid_dep_complete}: \code{TRUE} if \code{complete.cases(top, bot)}
\item \code{valid_dep_overlap}: \code{TRUE} if no gaps or overlaps in adjacent horizons
\item \code{valid_dep_all}: \code{TRUE} if all valid_ are \code{TRUE}
}
}
\description{
This function inspects a series of horizon depths for common errors:
\enumerate{
\item bottom depth is shallower than the top depth
\item equal top and bottom depths
\item missing top or bottom depths (e.g. \code{NA})
\item gap or overlap between adjacent horizons
}
}
\details{
This function peforms several validations list above to check horizon depths for errors. Compared to \code{checkHzDepthLogic}, it is slightly faster, reports more intuitive results (e.g. TRUE = pass), reports the results at the horizon level rather than the pedon, appends the results to the horizon table, exports the results \code{aqp.env}, and incorporates options to rmHzErrors and pad bottom horizon depths with +1cm.
}
\examples{

## sample data

data(sp3)

# these data should be clean
depths(sp3) <- id ~ top + bottom
sp3_val <- validate_depths(sp3)

head(sp3_val)

# sample data with errors
test <- data.frame(peiid  = 1:6, 
                   hzdept = c(0, 2, 10, 25, 50, 100), 
                   hzdepb = c(0, 11, 25, 50, 100, NA)
                   )

test_val <- validate_depths(test)

}
\author{
S.M. Roecker, D.E. Beaudette, A.G. Brown
}
