% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMunsell.R
\name{validateMunsell}
\alias{validateMunsell}
\title{Validate Standard Munsell Notation}
\usage{
validateMunsell(m, vRange = c(1, 12), cRange = c(0, 50))
}
\arguments{
\item{m}{character vector of colors in Munsell notation}

\item{vRange}{numeric vector of length 2, range of expected Munsell value}

\item{cRange}{numeric vector of length 2, range of expected Munsell chroma}
}
\value{
logical vector,
\itemize{
\item \code{TRUE}: valid Munsell notation
\item \code{FALSE}: invalid notation
}
}
\description{
This function validates Munsell color notation according to the following requirements:
\enumerate{
\item \emph{hue} must not be absent, and within the set of "standard hues", see \code{\link[=huePosition]{huePosition()}}
\item \emph{value} must not be absent, and within \code{vRange}
\item \emph{chroma} must be within \code{cRange}, and is only allowed to be absent in neutral colors (e.g. \verb{N 4/})
}
}
\examples{

# valid
validateMunsell('5Y 6/8')
validateMunsell('N 4/')
validateMunsell('2.5Y 4/0')

# invalid
validateMunsell('5G 0/4')
validateMunsell(NA)

# mixture
validateMunsell(c('5G 4/4', 'N 2/', NA, 'NA', '100R 3/3'))

}
\seealso{
\code{\link[=huePosition]{huePosition()}}
}
