% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{validSpatialData,SoilProfileCollection-method}
\alias{validSpatialData,SoilProfileCollection-method}
\alias{validSpatialData}
\title{Get names of columns in site table}
\usage{
\S4method{validSpatialData}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Are the contents of @sp valid: n x 2 matrix? If not, then contents of @sp in the SoilProfileCollection are an empty SpatialPoints object.
}
