% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauW.R
\name{xtableTauW}
\alias{xtableTauW}
\title{Format a LaTeX table with results}
\usage{
xtableTauW(result.tau, file.name = "tau_results_table.tex")
}
\arguments{
\item{result.tau}{results returned by \code{tauW}}

\item{file.name}{name of file to write output TeX file; Default: \code{file.name="tau_results_table.tex"}}
}
\description{
Format a LaTeX table with results
}
