\name{aqp-package}
\alias{aqp-package}
\alias{aqp}
\docType{package}
\title{
Algorithms for Quantitative Pedology
}
\description{
R package for working with and plotting soil profile data.
}


\author{Dylan E. Beaudette <debeaudette@ucdavis.edu>}
\references{http://casoilresource.lawr.ucdavis.edu/}
\keyword{ package }
\seealso{\code{\link{sp1}, \link{sp2}, \link{sp3}}}

