\name{aqp-package}
\alias{aqp-package}
\alias{aqp}
\docType{package}
\title{
Algorithms for Quantitative Pedology
}
\description{
R package for working with and plotting soil profile data.
}


Index:
\preformatted{
format_slices           Format Depth Slices of a Soil Profile
initProfile             Initialise a SoilProfile Object
initProfileList         Intitalize a SoilProfileList Object
make.segments           Low-Level Lattice Plotting of Line Segments
munsell                 Munsell-RGB Lookup Table for Common Soil Colors
munsell2rgb             Convert Munsell Notation to RGB
panel.depth_function    Lattice Panel Function for Soil Profiles
plot_slices             Map Depth Slices
prepanel.depth_function
                        Pre-Panel Function for panel.depth_function
print.SoilProfile       Print Method for SoilProfile and
                        SoilProfileList Class Objects
profile_compare         Numerical Soil Profile Comparison
profile_plot.default    Profile Plot
random_profile          Random Profile
soil.slot               Soil Profile Slotting
soil.slot.multiple      Multivariate Soil Profile Slotting
sp1                     Soil Profile Data Example 1
sp2                     Honcut Creek Soil Profile Data
sp3                     Soil Profile Data Example 3
unroll                  Unroll Genetic Horizons
}

\author{
Dylan Beaudette <debeaudette@ucdavis.edu>
}
\references{
http://casoilresource.lawr.ucdavis.edu/
}
\keyword{ package }
\seealso{\code{\link{sp1}, \link{sp2}, \link{sp3}}}
\examples{
# TODO 
}
