\name{aqp-package}
\alias{aqp-package}
\alias{aqp}
\docType{package}
\title{
Algorithms for Quantitative Pedology
}
\description{
R package for working with and plotting soil profile data.
}
\details{
\tabular{ll}{
Package: \tab aqp\cr
Version: \tab 0.75\cr
Date: \tab 2010-03-14\cr
Depends: \tab R (>= 2.8.0), lattice, cluster, sp, maptools\cr
Suggests: \tab RColorBrewer, plyr\cr
License: \tab GPL (>= 2)\cr
Packaged: \tab 2010-03-14 11:00:00 UTC; dylan\cr
Built: \tab R 2.10.1; ; 2010-03-23 18:40:20 UTC; unix\cr
}

Index:
\preformatted{
format_slices           Format Depth Slices of a Soil Profile
initProfile             Initialise a SoilProfile Object
initProfileList         Intitalize a SoilProfileList Object
make.segments           Low-Level Lattice Plotting of Line Segments
munsell                 Munsell-RGB Lookup Table for Common Soil Colors
munsell2rgb             Convert Munsell Notation to RGB
panel.depth_function    Lattice Panel Function for Soil Profiles
plot_slices             Map Depth Slices
prepanel.depth_function
                        Pre-Panel Function for panel.depth_function
print.SoilProfile       Print Method for SoilProfile and
                        SoilProfileList Class Objects
profile_compare         Numerical Soil Profile Comparison
profile_plot.default    Profile Plot
random_profile          Random Profile
soil.slot               Soil Profile Slotting
sp1                     Soil Profile Data Example 1
sp2                     Honcut Creek Soil Profile Data
sp3                     Soil Profile Data Example 3
unroll                  Unroll Genetic Horizons
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Dylan Beaudette <debeaudette@ucdavis.edu>

Maintainer: Dylan Beaudette <debeaudette@ucdavis.edu>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# TODO 
}
