% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSPC.R
\name{checkSPC}
\alias{checkSPC}
\title{Test for a valid SoilProfileCollection}
\usage{
checkSPC(x)
}
\arguments{
\item{x}{a \code{SoilProfileCollection} object}
}
\value{
TRUE or FALSE. Consider using \code{rebuildSPC()} if FALSE.
}
\description{
Test for a valid SoilProfileCollection
}
\details{
Test for valid \code{SoilProfileCollection} by checking for slots defined in
the class prototype. Likely only used between major versions of \code{aqp} where
internal structure of \code{SoilProfileCollection} has changed. Use
\code{checkHzDepthLogic} to check for common errors in horizon depths.
}
\seealso{
\code{\link{rebuildSPC}}, \code{\link{checkHzDepthLogic}}
}
\author{
D.E. Beaudette
}
