% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorChart.R
\name{colorChart}
\alias{colorChart}
\title{Visualize soil colors in Munsell notation according to within-group frequency.}
\usage{
colorChart(
  m,
  g = factor("All"),
  chip.cex = 3,
  alpha.wt = 10,
  alpha.toggle = TRUE,
  annotate = FALSE,
  annotate.cex = chip.cex * 0.25
)
}
\arguments{
\item{m}{character vector of color in Munsell notation ('10YR 4/6')}

\item{g}{factor describing group membership, typically a generalization of horizon designation, default value will generat a fake grouping that covers all of the colors in \code{m}}

\item{chip.cex}{scaling factor applied to each "chip"}

\item{alpha.wt}{weighted applied to chip transparency}

\item{alpha.toggle}{logical, automatically set full opacity (no transparency) when all chips share the same frequency}

\item{annotate}{logical, annotate chip frequency}

\item{annotate.cex}{scaling factor for chip frequency annotation}
}
\value{
a \code{trellis} object
}
\description{
Visualize soil colors in Munsell notation according to within-group frequency.
}
\examples{

# required for latticeExtra:useOuterStrips
if(!requireNamespace('latticeExtra')) {
  
  # two hue pages
  ric <- expand.grid(
    hue = c('5YR', '7.5YR'),
    value = 2:8,
    chroma = 2:8
  )
  
  # combine hue, value, chroma into standard Munsell notation
  ric <- sprintf("\%s \%s/\%s", ric$hue, ric$value, ric$chroma)
  
  # note that chip frequency-based transparency is disabled 
  # because all chips have equal frequency
  colorChart(ric, chip.cex = 3)
  
  # annotation of frequency
  colorChart(ric, chip.cex = 3, annotate = TRUE)
  
  # disable auto-toggling of transparency when all chips equal
  colorChart(ric, chip.cex = 3, alpha.toggle = FALSE)
  
  
  # bootstrap to larger size
  ric.big <- sample(ric, size = 100, replace = TRUE)
  
  # note that chip frequency is encoded in transparency
  colorChart(ric.big, chip.cex = 3)
  colorChart(ric.big, chip.cex = 3, annotate = TRUE)
  
  # adjust transparency weighting
  colorChart(ric.big, chip.cex = 3, annotate = TRUE, alpha.wt = 20)
  
  
  # simulate colors based dE00 thresholding
  p <- list(
    list(m = '10YR 4/4', thresh = 10, hues = c('10YR', '7.5YR'))
  )
  
  # perform 500 simulations
  s <- simulateColor(method = 'dE00', n = 500, parameters = p)
  
  # result is a list, use the first element
  colorChart(s[[1]], chip.cex = 3, alpha.wt = 10, annotate = TRUE)
  
}

}
