% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R,
%   R/SoilProfileCollection-spatial.R
\docType{methods}
\name{coordinates,SoilProfileCollection-method}
\alias{coordinates,SoilProfileCollection-method}
\alias{coordinates}
\alias{coordinates<-,SoilProfileCollection,formula-method}
\alias{coordinates<-,SoilProfileCollection,character-method}
\title{Get Soil Profile Coordinates}
\usage{
\S4method{coordinates}{SoilProfileCollection}(obj)

\S4method{coordinates}{SoilProfileCollection,formula}(object) <- value

\S4method{coordinates}{SoilProfileCollection,character}(object) <- value
}
\arguments{
\item{obj}{a SoilProfileCollection}

\item{object}{A SoilProfileCollection}

\item{value}{A formula specifying columns containing x and y coordinates, or character with the column names}
}
\description{
Get coordinates of each profile
}
\examples{

data(sp5)

# coordinates are stored in x and y column of site
sp5$x <- rnorm(length(sp5))
sp5$y <- rnorm(length(sp5))

# coordinates takes a formula object as input
coordinates(sp5) <- ~ x + y

}
