% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-setters.R
\name{depths<-,SoilProfileCollection-method}
\alias{depths<-,SoilProfileCollection-method}
\alias{depths<-}
\alias{depths<-,data.frame-method}
\title{Initialize a SoilProfileCollection from a data.frame object}
\usage{
\S4method{depths}{SoilProfileCollection}(object) <- value

\S4method{depths}{data.frame}(object) <- value
}
\arguments{
\item{object}{An object to promote to SoilProfileCollection (inherits from data.frame)}

\item{value}{A formula specifying the unique profile ID, top and bottom depth column names}
}
\description{
Initialize a SoilProfileCollection from a data.frame object
}
\details{
The input horizon data, and the resulting profile order, is sorted based on unique profile ID and top depth. ID columns are converted to character, depth columns are converted to integer. If \code{NA} values exist in all of the top depths, a prototype with 1 horizon per profile ID is returned, with \code{NA} in all non-essential columns. If the input \code{object} has 0 rows, a prototype with 0 horizons and 0 rows, but same column names as \code{object}, is returned.
}
\examples{
## init SoilProfileCollection objects from data.frame of horizon data

# load demo data
data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# plot
plot(sp1)

# number of profiles
length(sp1)

# number of horizons
nrow(sp1)

}
