% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{diagnostic_hz}
\alias{diagnostic_hz}
\alias{diagnostic_hz,SoilProfileCollection-method}
\title{Retrieve diagnostic data from SoilProfileCollection}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get diatnostic data from SoilProfileCollection. Result is returned in the same \code{data.frame} class used to initially construct the SoilProfileCollection.
}
