% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{diagnostic_hz,SoilProfileCollection-method}
\alias{diagnostic_hz,SoilProfileCollection-method}
\alias{diagnostic_hz}
\title{Retrieve diagnostic data from SoilProfileCollection}
\usage{
\S4method{diagnostic_hz}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get diagnostic feature data from SoilProfileCollection. Result is returned in the same \code{data.frame} class used to initially construct the SoilProfileCollection.
}
