% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,SoilProfileCollection-method,$<-,SoilProfileCollection-method}
\title{Access column of horizon or site data in a SoilProfileCollection}
\usage{
\S4method{$}{SoilProfileCollection}(x, name)
}
\arguments{
\item{name}{a single column name in site or horizon table}

\item{object}{a SoilProfileCollection}
}
\description{
Get/set the data from a column accessed by name \code{spc$name}. Horizon data takes precedence. Column names are not shared between site and horizons. 

When using \code{$<-}, the length of input and output matching either the number of sites or number of horizons is used to determine which slot new columns are assigned to.
}
\examples{

data(sp1)

depths(sp1) <- id ~ top + bottom

# get data from a column by name (prop)
sp1$prop

}
