% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-operators.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,SoilProfileCollection-method}
\title{Get data from column of horizon or site data in a SoilProfileCollection}
\usage{
\S4method{$}{SoilProfileCollection}(x, name)
}
\arguments{
\item{x}{a SoilProfileCollection}

\item{name}{a single column name in site or horizon table}
}
\description{
Get the data from a column accessed by name \code{x$name}. Column names other than profile ID are not shared between site and horizons.
}
\examples{

data(sp1)

depths(sp1) <- id ~ top + bottom

# get data from a column by name (prop)
sp1$prop

}
