% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainPlotSPC.R
\name{explainPlotSPC}
\alias{explainPlotSPC}
\title{Visual Explanation for \code{plotSPC}}
\usage{
explainPlotSPC(x, ...)
}
\arguments{
\item{x}{a \code{SoilProfilecollection} object}

\item{\dots}{arguments passed to \code{\link{plotSPC}}}
}
\value{
a list of internally-used ordering vectors and graphical offsets / scaling factors
}
\description{
Create a visual explanation for the many arguments to \code{plotSPC}. Call this function instead of \code{plotSPC}, all objects after \code{x} are passed on to \code{plotSPC}. Nearly all of the figures in the \href{https://ncss-tech.github.io/AQP/aqp/aqp-intro.html}{Introduction to SoilProfileCollection Objects tutorial} are created with this function.
}
\seealso{
\code{\link{plotSPC}}
}
\author{
D.E. Beaudette
}
\keyword{manip}
