% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillHzGaps.R
\name{fillHzGaps}
\alias{fillHzGaps}
\title{Find and Fill Horizon Gaps}
\usage{
fillHzGaps(x, flag = FALSE)
}
\arguments{
\item{x}{\code{SoilProfileCollection} object}

\item{flag}{logical, flag empty horizons that have been added}
}
\value{
\code{SoilProfileCollection} object
}
\description{
This function attempt to find "gaps" in the horizon records of a \code{SoilProfileCollection} object and fill with place holder horizons.

Gaps are defined as:
\itemize{
\item within each profile, for horizons \code{i} to \code{n_hz}:
\item \verb{bottom_i != top_i+1 (but only to i = 1:(n_hz - 1)}
}
}
\examples{

data(sp4)
depths(sp4) <- id ~ top + bottom

# introduce gaps
idx <- c(2, 8, 12)
sp4$top[idx] <- NA

# check
horizons(sp4)[idx, ]

# remove problematic horizons
x <- HzDepthLogicSubset(sp4, byhz = TRUE)

# gaps and now problematic profiles
par(mar = c(0, 0, 0, 1))
plotSPC(x, width = 0.3, default.color = 'royalblue')

z <- fillHzGaps(x, flag = TRUE)

# BUG: plotSPC can't use logical data for color
z$.filledGap <- as.factor(z$.filledGap)
plotSPC(z, width = 0.3, color = '.filledGap', show.legend = FALSE)

}
