% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqp-label-placement-solvers.R
\name{fixOverlap}
\alias{fixOverlap}
\title{Fix Overlap within a Sequence}
\usage{
fixOverlap(x, thresh = 0.6, method = c("S", "E"), trace = FALSE, ...)
}
\arguments{
\item{x}{vector of initial positions, pre-sorted}

\item{thresh}{numeric, overlap threshold defined on the same scale as \code{x}}

\item{method}{character vector, 'S' for simulated annealing via \code{\link[=SANN_1D]{SANN_1D()}} or 'E' for electrostatic simulation via \code{\link[=electroStatics_1D]{electroStatics_1D()}}}

\item{trace}{logical, return full output}

\item{\dots}{additional arguments to \code{\link[=SANN_1D]{SANN_1D()}} or \code{\link[=electroStatics_1D]{electroStatics_1D()}}}
}
\value{
When \code{trace = FALSE}, a vector of the same length as \code{x}, preserving rank-ordering and boundary conditions. When \code{trace = TRUE} a list containing the new sequence along with information about objective functions and decisions made during adjustment of \code{x}.
}
\description{
Fix Overlap within a Sequence
}
\examples{

s <- c(1, 2, 2.3, 4, 5, 5, 7)

# simulated annealing, solution is non-deterministic
fixOverlap(s, thresh = 0.6, method = 'S')

# electrostatics-inspired simulation of particles
# solution is deterministic
fixOverlap(s, thresh = 0.6, method = 'E')

}
\seealso{
\code{\link[=electroStatics_1D]{electroStatics_1D()}}, \code{\link[=SANN_1D]{SANN_1D()}}
}
