% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap-fixOverlap.R
\name{fixOverlap}
\alias{fixOverlap}
\title{Fix Overlap within a Sequence}
\usage{
fixOverlap(
  x,
  thresh = 0.6,
  adj = thresh/2,
  min.x = 0.8,
  max.x = max(x) + 0.2,
  coolingRate = 0.95,
  maxIter = 1000,
  trace = FALSE
)
}
\arguments{
\item{x}{vector of horizontal positions}

\item{thresh}{horizontal threshold defining "overlap", typically < 1, ideal values likely in (0.3, 0.8)}

\item{adj}{adjustments are tested within \code{runif(min = adj * -1, max = adj)}}

\item{min.x}{left-side boundary condition}

\item{max.x}{right-side boundary condition}

\item{coolingRate}{rate at which \code{adj} is decreased after a successful iteration}

\item{maxIter}{maximum number of iterations to attempt before giving up and returning a regularly-spaced sequence}

\item{trace}{print diagnostics, result is a \code{list}}
}
\value{
When \code{trace = FALSE}, a vector of the same length as \code{x}, preserving rank-ordering and boundary conditions. When \code{trace = TRUE} a list containing the new sequence along with the number of overlapping elements at each iteration.
}
\description{
This function attempts to iteratively adjust a sequence until values are no longer within a given threshold, until \code{maxIter} is reached.
}
\examples{

x <- c(1, 2, 3, 3.4, 3.5, 5, 6, 10)

fixOverlap(x, thresh = 0.2, max.x = 10, trace = TRUE)

}
\author{
D.E. Beaudette
}
