% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentClasses.R
\name{fragmentClasses}
\alias{fragmentClasses}
\title{Title}
\usage{
fragmentClasses(
  sys = c("usda_simplified", "usda", "international", "unified", "aashto",
    "mod.wentworth"),
  flat = FALSE,
  rounded = FALSE
)
}
\arguments{
\item{sys}{}

\item{flat}{}

\item{rounded}{}
}
\value{
named vector of fragment diameters
}
\description{
Title
}
\examples{

fragmentClasses()
fragmentClasses(flat = TRUE)

fragmentClasses(sys = 'usda')
fragmentClasses(sys = 'USDA', flat = TRUE)

fragmentClasses(sys = 'international')

fragmentClasses(sys = 'unified')

fragmentClasses(sys = 'aashto')
fragmentClasses(sys = 'aashto', rounded = TRUE)

fragmentClasses(sys = 'mod.wentworth')

}
\references{
Schoeneberger, P.J., D.A. Wysocki, E.C. Benham, and Soil Survey Staff. 2012. Field book for describing and sampling soils, Version 3.0. Natural Resources Conservation Service, National Soil Survey Center, Lincoln, NE.
}
