% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.hz.R
\name{generalize.hz}
\alias{generalize.hz}
\alias{generalizeHz,character-method}
\alias{generalizeHz,SoilProfileCollection-method}
\title{Generalize Horizon Names}
\usage{
generalize.hz(
  x,
  new,
  pattern,
  non.matching.code = "not-used",
  hzdepm = NULL,
  ordered = !missing(hzdepm),
  ...
)

\S4method{generalizeHz}{character}(
  x,
  new,
  pattern,
  non.matching.code = "not-used",
  hzdepm = NULL,
  ordered = !missing(hzdepm),
  ...
)

\S4method{generalizeHz}{SoilProfileCollection}(
  x,
  new,
  pattern,
  non.matching.code = "not-used",
  hzdepm = NULL,
  ordered = !missing(hzdepm),
  gendesgnname = "genhz",
  ...
)
}
\arguments{
\item{x}{a character vector of horizon names or a \code{SoilProfileCollection}}

\item{new}{a character vector of new horizon classes}

\item{pattern}{a character vector of REGEX patterns, same length as \code{new}}

\item{non.matching.code}{label used for any horizon not matching any \code{pattern}}

\item{hzdepm}{a numeric vector of horizon mid-points; \code{NA} values in \code{hzdepm} will result in \code{non.matching.code} (or \code{NA} if not defined) in result}

\item{ordered}{by default, the result is an ordered factor when \code{hzdepm} is defined.}

\item{...}{additional arguments passed to \code{grep()} such as \code{perl=TRUE} for advanced REGEX}

\item{gendesgnname}{Generalized Horizon Designation column name (to be created/updated when \code{x} is a \code{SoilProfileCollection})}
}
\value{
(ordered) factor of the same length as \code{x} (if character) or as number of horizons in \code{x} (if SoilProfileCollection)
}
\description{
Generalize a vector of horizon names, based on new classes, and REGEX
patterns. Or create a new column \code{gendesgnname} in a SoilProfileCollection (requires a horizon designation name to be defined for the collection, see details)
}
\details{
When \code{x} is a SoilProfileCollection the \code{gendesgnname} column will be updated with the factor results. This requires that the "horizon designation name" metadata be defined for the collection to set the column for input designations. See \code{hzdesgnname()}.
}
\examples{

data(sp1)

# check original distribution of hz designations
table(sp1$name)

# generalize
sp1$genhz <- generalize.hz(sp1$name,
                           new=c('O','A','B','C','R'),
                           pattern=c('O', '^A','^B','C','R'))

# see how we did / what we missed
table(sp1$genhz, sp1$name)


## a more advanced example, requries perl=TRUE
# example data
x <- c('A', 'AC', 'Bt1', '^AC', 'C', 'BC', 'CB')

# new labels
n <- c('A', '^AC', 'C')
# patterns:
# "A anywhere in the name"
# "literal '^A' anywhere in the name"
# "C anywhere in name, but without preceding A"
p <- c('A', '^A', '(?<!A)C')

# note additional argument
res <- generalize.hz(x, new = n, pattern=p, perl=TRUE)

# double-check: OK
table(res, x)


}
\seealso{
\code{hzdesgnname()}
}
\author{
D.E. Beaudette
}
\keyword{manip}
