% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize.hz.R
\name{genhzTableToAdjMat}
\alias{genhzTableToAdjMat}
\title{Convert cross-tabulation to adjacency matrix.}
\usage{
genhzTableToAdjMat(tab)
}
\arguments{
\item{tab}{table, cross-tabulation of original and generalized horizon labels e.g. \code{table(original, genhz)}}
}
\value{
matrix of numeric values
}
\description{
Convert a cross-tabulation of e.g. original horizon designations vs. generalized horizon labels to adjacency matrix form.
}
\author{
D.E. Beaudette
}
