% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSPC.R
\name{groupSPC}
\alias{groupSPC}
\title{Store groupings within a profile collection.}
\usage{
groupSPC(object, ...)
}
\arguments{
\item{object}{SoilProfileCollection.}

\item{...}{One or more expressions evaluated within the context of \code{object} that resolve to vectors that can be coerced to factor "groups."}
}
\description{
Store groupings within a profile collection.
}
