% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growEmptyHz.R
\name{growEmptyHz}
\alias{growEmptyHz}
\title{Grow a \code{SoilProfileCollection} with Empty Horizons}
\usage{
growEmptyHz(x, z, direction = c("down", "up"))
}
\arguments{
\item{x}{a \code{SoilProfilecollection} object}

\item{z}{new top or bottom depth (see \code{direction}), after growing with empty horizons}

\item{direction}{empty horizons are "grown" \emph{down} to \code{z} or \emph{up} to \code{z}}
}
\value{
a \code{SoilProfilecollection} object
}
\description{
This function appends a single "empty" horizon to each profile that is shallower than \code{z}. Horizon IDs are reset and the horizons are re-sorted according to profile ID and horizon top depth. This will fail if horizon top depths are \code{NA}. Adding empty horizons to profiles is one way to ensure that data (even NA) will be returned from each profile when sliced at depths exceeding some profiles in the collection.
}
\examples{

data(sp4)
depths(sp4) <- id ~ top + bottom

par(mar = c(0, 0, 0, 0))

plotSPC(sp4)

g <- growEmptyHz(sp4, 50)
plotSPC(g, color = 'hzID', show.legend = FALSE)
plotSPC(g, color = 'Ca', show.legend = FALSE)

d <- dice(g, fm = 0:50 ~ .)
plotSPC(d, color = 'sliceID', show.legend = FALSE)
plotSPC(d, color = 'Ca', show.legend = FALSE)

}
\author{
D.E. Beaudette
}
