% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessColumnNames.R
\name{guessHzTexClName}
\alias{guessHzTexClName}
\title{Guess Horizon Texture Class Column Name}
\usage{
guessHzTexClName(x)
}
\arguments{
\item{x}{A SoilProfileCollection}
}
\value{
Character containing horizon texture class column name.
}
\description{
This function is used to provide texture class attribute column name to functions that use it to determine taxonomic criteria. It will use regular expressions to match \code{'texcl'} which is typically the texture of the fine earth fraction, without modifiers or in-lieu textures. Alternately, it will match \code{'texture'} for cases where \code{'texcl'} is absent (e.g. in NASIS components).
}
\examples{

a <- data.frame(id = 1, top = c(0,10), bottom=c(10,40), texture=c("A","Bw"))
depths(a) <- id ~ top + bottom

# store guess in metadata
hzdesgnname(a) <- guessHzTexClName(a)

# inspect result
hzdesgnname(a)

}
\author{
Andrew G. Brown
}
