% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{horizonDepths<-}
\alias{horizonDepths<-}
\alias{horizonDepths<-,SoilProfileCollection-method}
\alias{horizonDepths,SoilProfileCollection-method}
\alias{horizonDepths}
\title{Set horizon depth column names}
\usage{
horizonDepths(object) <- value

\S4method{horizonDepths}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a character vector of length two with names of columns containing numeric top and bottom depths}
}
\description{
Set column name containing horizon ID

Get column names containing horizon depths
}
