% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{horizonNames}
\alias{horizonNames}
\alias{horizonNames,SoilProfileCollection-method}
\alias{horizonNames<-}
\alias{horizonNames<-,SoilProfileCollection-method}
\title{Get names of columns in horizon table}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of columns in horizons \code{length(horizonNames(object))}}
}
\description{
Get names of columns in horizon table.

Set horizon column names
}
