% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{horizonNames<-}
\alias{horizonNames<-}
\alias{horizonNames<-,SoilProfileCollection-method}
\alias{horizonNames,SoilProfileCollection-method}
\alias{horizonNames}
\title{Set horizon column names}
\usage{
\S4method{horizonNames}{SoilProfileCollection}(object) <- value

\S4method{horizonNames}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of columns in horizons \code{length(horizonNames(object))}}
}
\description{
Set horizon column names

Get names of columns in horizon table.
}
