% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{hz_lag}
\alias{hz_lag}
\title{Find lagged horizon values}
\usage{
hz_lag(
  object,
  vars,
  lag = 1,
  idcol = "id",
  depthcols = c("top", "bottom"),
  order = FALSE
)
}
\arguments{
\item{object}{a \code{data.frame}}

\item{vars}{character: column names, to lag.}

\item{lag}{integer: number of horizons to lag}

\item{idcol}{character: column name of the pedon ID within the object.}

\item{depthcols}{a character vector of length 2 specifying the names of the horizon depths (e.g. \code{c("top", "bottom")}).}

\item{order}{logical: indicating whether or not to order the #'}
}
\value{
A \code{data.frame} with lagged values.
}
\description{
This function finds adjacent values to a horizon values at lagged distances.
}
\details{
.
}
\examples{

h <- data.frame(
id = 1,
top    = c(0,  25, 44, 46, 50),
bottom = c(25, 44, 46, 50, 100),
texcl     = c("SL", "SL", "CL", "CL", "L"),
clay   = c(10, 12, 27, 35, 16)
)

h |> hz_lag(c("texcl", "clay"), 1)

h |> 
hz_lag(c("texcl", "clay"), 1) |> 
cbind(h) |>
transform(
clay_dif = clay - clay_lag.1,
texcl_contrast = paste0(texcl, "-", texcl_lag.1))

}
\author{
Stephen Roecker
}
