% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HzDepthLogicSubset.R
\name{HzDepthLogicSubset}
\alias{HzDepthLogicSubset}
\title{Subset \code{SoilProfileCollection} Objects or Horizons via \code{checkHzDepthLogic}}
\usage{
HzDepthLogicSubset(x, byhz = FALSE)
}
\arguments{
\item{x}{a \code{SoilProfileCollection} object}

\item{byhz}{logical, evaluate horizon depth logic at the horizon level (profile level if \code{FALSE})}
}
\value{
a \code{SoilProfileCollection} object
}
\description{
Subset \code{SoilProfileCollection} Objects or Horizons via \code{checkHzDepthLogic}
}
