% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzDesgn,SoilProfileCollection-method}
\alias{hzDesgn,SoilProfileCollection-method}
\alias{hzDesgn}
\title{Get horizon designation column name}
\usage{
\S4method{hzDesgn}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get horizon designation names
}
