% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{hzdesgnname}
\alias{hzdesgnname}
\alias{hzdesgnname,SoilProfileCollection-method}
\alias{hzdesgname<-}
\title{Get horizon designation column name}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{character, name of column containing horizon designations}
}
\description{
Get column name containing horizon designation name

Set horizon designation column name
}
\examples{

data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# set horizon designation column
hzdesgnname(sp1) <- "name"

# get horizon designation column
hzdesgnname(sp1)
}
