% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzID}
\alias{hzID}
\alias{hzID,SoilProfileCollection-method}
\title{Get horizon IDs}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get vector containing horizon IDs
}
