% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R, R/setters.R
\docType{methods}
\name{hzID}
\alias{hzID}
\alias{hzID,SoilProfileCollection-method,hzID<-,SoilProfileCollection-method}
\alias{hzID<-}
\alias{hzID<-,SoilProfileCollection-method}
\title{Get/set horizon IDs}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of horizons \code{nrow(object)}}
}
\description{
Get or set vector containing horizon IDs

Set vector containing horizon IDs
}
