% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hzidname<-}
\alias{hzidname<-}
\alias{hzidname<-,SoilProfileCollection-method}
\alias{hzidname,SoilProfileCollection-method}
\alias{hzidname}
\title{Set horizon ID column name}
\usage{
hzidname(object) <- value

\S4method{hzidname}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{character, column name containing unique horizon ID values}
}
\description{
Set unique horizon ID column name

Get column name containing unique horizon ID
}
\examples{
data(sp1)

# promote to SPC
depths(sp1) <- id ~ top + bottom

# create new horizon ID
sp1$hzIDrev <- rev(sp1$hzID)

# set horizon designation column
hzidname(sp1) <- "hzIDrev"

# get horizon designation column
hzidname(sp1)

}
