% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hztexclname}
\alias{hztexclname}
\alias{hztexclname,SoilProfileCollection-method}
\title{Get horizon texture class column name}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get column name containing horizon designation name
}
