% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{hztexclname,SoilProfileCollection-method}
\alias{hztexclname,SoilProfileCollection-method}
\alias{hztexclname}
\title{Get horizon texture class column name}
\usage{
\S4method{hztexclname}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get column name containing horizon designation name
}
