% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{idname,SoilProfileCollection-method}
\alias{idname,SoilProfileCollection-method}
\alias{idname}
\title{Get profile ID column name}
\usage{
\S4method{idname}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get column name containing unique profile IDs
}
