% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertLabelColor.R
\name{invertLabelColor}
\alias{invertLabelColor}
\title{Make High Contrast Label Colors}
\usage{
invertLabelColor(colors, threshold = 0.6)
}
\arguments{
\item{colors}{vector of colors}

\item{threshold}{black | white threshold}
}
\value{
vector of label colors
}
\description{
Generate a vector of white or black label colors conditioned on a vector of colors to maximize label contrast.
}
\author{
D.E. Beaudette
}
