% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_profiles.R
\name{L1_profiles}
\alias{L1_profiles}
\title{Create Representative Soil Profiles via L1 Estimator}
\usage{
L1_profiles(
  x,
  fm,
  basis = 1,
  method = c("regex", "simple", "constant"),
  maxDepthRule = c("max", "min"),
  maxDepthConstant = NULL,
  strict = FALSE
)
}
\arguments{
\item{x}{\code{SoilProfileCollection} object}

\item{fm}{formula, for example: "group ~ p1 + p2 + p3", where "group" is a site-level grouping variable, and "p1", "p2", and "p3" are horizon level variables}

\item{basis}{positive integer, aggregation basis (e.g. 1 for 1-depth-unit intervals). Values other than 1 are not currently supported.}

\item{method}{soil depth evaluation method: "regex" for regular expression, "simple", or "constant". See details.}

\item{maxDepthRule}{maximum depth rule: "max" or "min" See details.}

\item{maxDepthConstant}{positive integer, maximum depth when \code{maxDepthRule = 'constant'}}

\item{strict}{passed to \code{slice}}
}
\value{
a \code{SoilProfileCollection} object
}
\description{
The L1 estimator, or \href{https://en.wikipedia.org/wiki/Geometric_median}{geometric median}, is a multivariate generalization of the (univariate) \href{https://en.wikipedia.org/wiki/Median}{median} concept. This function performs a multivariate aggregation (via L1 estimator) according to a suite of ratio-scale soil properties. The L1 estimator is applied to soil profile data that have been sliced to a 1-depth-unit basis.

See the \href{https://ncss-tech.github.io/AQP/aqp/L1-profiles.html}{L1 Profiles Tutorial} for additional examples.
}
\note{
This function requires the `Gmedian` package.
}
\references{
Cardot, H., Cenac, P. and Zitt, P-A. (2013). Efficient and fast estimation of the geometric median in Hilbert spaces with an averaged stochastic gradient algorithm. Bernoulli, 19, 18-43.
}
