% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launderMunsell.R
\name{launderMunsell}
\alias{launderMunsell}
\title{Fix Common Errors in Munsell Notation}
\usage{
launderMunsell(m, verbose = FALSE)
}
\arguments{
\item{m}{character vector of Munsell colors}

\item{verbose}{logical, optionally return a \code{data.frame} comparing modifications}
}
\value{
either character vector, or when \code{verbose = TRUE} a \code{data.frame}
}
\description{
This function will "launder" colors in Munsell notation, fixing (and optionally flagging) the following:
\itemize{
\item colors specified with 0-chroma will be converted to neutral hue (N) and original value
\item for neutral colors, any chroma >0 will be set to 0
}

Examples:
\itemize{
\item '10YR 2/0' -> 'N 2/0'
\item 'N 4/1 -> 'N 4/0'
}
}
\examples{

# will be converted to 'N 4/0'
launderMunsell('5G 4/0')

}
