% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,SoilProfileCollection-method}
\title{Get the number of profiles in a SoilProfileCollection}
\usage{
\S4method{length}{SoilProfileCollection}(x)
}
\arguments{
\item{x}{a SoilProfileCollection}
}
\description{
Get the number of profiles in a SoilProfileCollection
}
