% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{max}
\alias{max}
\alias{max,SoilProfileCollection-method}
\title{Get the maximum bottom depth in a SoilProfileCollection}
\usage{
\S4method{max}{SoilProfileCollection}(x, v = NULL)
}
\arguments{
\item{x}{a SoilProfileCollection}

\item{v}{optional: horizon-level column name to refine calculation}
}
\description{
Get the deepest depth of description out of all profiles in a SoilProfileCollection. Data missing one or more of: bottom depth, profile ID, or any optional attribute are omitted using \code{complete.cases}.
}
