% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixMunsell.R
\name{mixMunsell}
\alias{mixMunsell}
\title{Mix Munsell Colors via Spectral Library}
\usage{
mixMunsell(
  x,
  w = rep(1, times = length(x))/length(x),
  mixingMethod = c("spectra", "estimate", "adaptive"),
  n = 1,
  keepMixedSpec = FALSE
)
}
\arguments{
\item{x}{vector of colors in Munsell notation}

\item{w}{vector of proportions, can sum to any number}

\item{mixingMethod}{approach used to simulate a mixture:
\itemize{
\item \code{spectra}: simulate a subtractive mixture of pigments, limited to available reference spectra
\item \code{estimate}: closest Munsell chip to a weighted mean of CIELAB coordinates
\item \code{adaptive}: use reference spectra when possible, falling-back to weighted mean of CIELAB coordinates
}}

\item{n}{number of closest matching color chips (\code{mixingMethod = spectra} only)}

\item{keepMixedSpec}{keep weighted geometric mean spectra, final result is a \code{list} (\code{mixingMethod = spectra} only)}
}
\value{
A \code{data.frame} with the closest matching Munsell color(s), unless \code{keepMixedSpec = TRUE} then a \code{list}.
}
\description{
Simulate mixing of colors in Munsell notation, similar to the way in which mixtures of pigments operate.
}
\details{
Pending
}
\references{
\itemize{
\item{inspiration / calculations based on the work of Scott Burns: }{\url{https://arxiv.org/ftp/arxiv/papers/1710/1710.06364.pdf}}

\item{related discussion on Stack Overflow: }{\url{https://stackoverflow.com/questions/10254022/implementing-kubelka-munk-like-krita-to-mix-colours-color-like-paint/29967630#29967630}}

\item{spectral library source: }{\url{https://www.munsellcolourscienceforpainters.com/MunsellResources/SpectralReflectancesOf2007MunsellBookOfColorGlossy.txt}}

}
}
\seealso{
\code{\link{munsell.spectra}}
}
\author{
D.E. Beaudette
}
