% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{names,SoilProfileCollection-method}
\alias{names,SoilProfileCollection-method}
\alias{names}
\title{Get names of columns in site and horizons table}
\usage{
\S4method{names}{SoilProfileCollection}(x)
}
\arguments{
\item{x}{a SoilProfileCollection}
}
\description{
Get names of columns in site and horizons table of a SoilProfileCollection.
}
