% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{nrow,SoilProfileCollection-method}
\title{Get the number of horizons in a SoilProfileCollection}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get the number of horizons in a SoilProfileCollection
}
