% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap-fixOverlap.R
\name{overlapMetrics}
\alias{overlapMetrics}
\title{Find and Quantify Overlap within a 1D Sequence}
\usage{
overlapMetrics(x, thresh)
}
\arguments{
\item{x}{vector of relative horizontal positions, one for each profile}

\item{thresh}{threshold defining "overlap", typically < 1, ideal values likely in (0.3, 0.8)

@return a \code{list}:
\itemize{
\item \code{idx}: index to overlapping elements in \code{x}
\item \code{ov}: total overlap (see details)
}}
}
\description{
Desc.
}
