% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{pbindlist}
\alias{pbindlist}
\title{Combine a list of SoilProfileCollection Objects}
\usage{
pbindlist(l)
}
\arguments{
\item{l}{a list of SoilProfileCollection objects}
}
\value{
a SoilProfileCollection object
}
\description{
Combine a list of SoilProfileCollection Objects
}
\details{
Input data must share a common depth unit, and if spatial data are present, a common CRS and coordinate names. In the case of non-conformal @idname and/or @depthcols, the first SoilProfileCollection is used as a template. Non-conforming spatial data are dropped from the final result.
}
\examples{
# example data
data(sp2, package = 'aqp')
depths(sp2) <- id ~ top + bottom
site(sp2) <- ~ surface

# copy pieces
x <- sp2[1:5, ]
y <- sp2[6:10, ]

# reset IDs and combine
profile_id(y) <- sprintf("\%s-copy", profile_id(y))

# this should work
z <- pbindlist(list(x, y))

# check
plot(z)
}
\author{
D.E. Beaudette and A.G. Brown
}
