% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{profile_id<-}
\alias{profile_id<-}
\alias{profile_id<-,SoilProfileCollection-method}
\alias{profile_id,SoilProfileCollection-method}
\alias{profile_id}
\title{Set profile IDs}
\usage{
\S4method{profile_id}{SoilProfileCollection}(object) <- value

\S4method{profile_id}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}

\item{value}{a unique vector of equal length to number of profiles \code{length(object)}}
}
\description{
Set vector containing profile IDs

Get or set a vector of profile IDs
}
