% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{restrictions,SoilProfileCollection-method}
\alias{restrictions,SoilProfileCollection-method}
\alias{restrictions}
\title{Retrieve restriction data from SoilProfileCollection}
\usage{
\S4method{restrictions}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\description{
Get restriction data from SoilProfileCollection. Result is returned in the same \code{data.frame} class used to initially construct the SoilProfileCollection.
}
