% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve.R
\name{sieve}
\alias{sieve}
\title{Sieve the Coarse Fraction of Soil}
\usage{
sieve(
  diameter,
  sieves = NULL,
  ordered = FALSE,
  prefix = "",
  new_names = NULL,
  ...
)
}
\arguments{
\item{diameter}{numeric. Vector of diameters of coarse fragments to "sieve". Default \code{sieves} are specified in millimeters.}

\item{sieves}{leave as \code{NULL} to use sieve names and fragment diameters defined by \code{\link[=fragmentClasses]{fragmentClasses()}}, or a named vector of fragment diameters. See examples.}

\item{ordered}{logical. Return as an ordered factor.}

\item{prefix}{character. Add a prefix to result names? Default: \code{""} adds no prefix. For example \code{"para"} might be used for size classes of pararock fragments.}

\item{new_names}{Optional: apply new labels to result classes. Should match length of \code{sieves}.}

\item{...}{additional arguments to \code{\link[=fragmentClasses]{fragmentClasses()}}, such as \code{sys}, \code{flat}, and \code{rounded}, see Details.}
}
\value{
character. Size class labels based on names of \code{sieves}, \code{new_names}, and \code{prefix} (if specified).
}
\description{
Sieve applies thresholds to a numeric vector of fragment diameter values, returning fragment size classes. Particle diameter thresholds are evaluated as \code{d < threshold}.
}
\examples{

# default
 
}
\references{
Soil Science Division Staff. 2017. Soil survey manual. C. Ditzler, K. Scheffe, and H.C. Monger (eds.). USDA Handbook 18. Government Printing Office, Washington, D.C.
}
