% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SoilProfileCollection.R
\docType{methods}
\name{siteNames<-}
\alias{siteNames<-}
\alias{siteNames<-,SoilProfileCollection-method}
\alias{siteNames}
\alias{siteNames,SoilProfileCollection-method}
\title{Set site column names}
\arguments{
\item{value}{a unique vector of equal length to number of columns in site: \code{length(siteNames(object))}}

\item{object}{a SoilProfileCollection}
}
\description{
Set site column names

Get names of columns in site table.
}
