% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilColorSignature.R
\name{soilColorSignature}
\alias{soilColorSignature}
\title{Soil Profile Color Signatures}
\usage{
soilColorSignature(
  spc,
  color,
  space = c("sRGB", "CIELAB"),
  r = NULL,
  g = NULL,
  b = NULL,
  method = c("colorBucket", "depthSlices", "pam"),
  pam.k = 3,
  RescaleLightnessBy = 1,
  useProportions = TRUE,
  pigmentNames = c(".white.pigment", ".red.pigment", ".green.pigment", ".yellow.pigment",
    ".blue.pigment"),
  apply.fun = lapply
)
}
\arguments{
\item{spc}{a \code{SoilProfileCollection} object}

\item{color}{horizon-level attributes, either character of length 1 specifiying a column containing Munsell or sRGB in hex notation, or character vector of three column names containing either sRGB or CIELAB color coordinates. sRGB color coordinates should be within the range of 0 to 1.}

\item{space}{character, either 'sRGB' or 'LAB', specifying color space}

\item{r}{deprecated, use \code{color} argument}

\item{g}{deprecated, use \code{color} argument}

\item{b}{deprecated, use \code{color} argument}

\item{method}{algorithm used to compute color signature, \code{colorBucket}, \code{depthSlices}, or \code{pam}}

\item{pam.k}{number of classes to request from \code{cluster::pam()}}

\item{RescaleLightnessBy}{rescaling factor for CIE LAB L-coordinate, ignored for \code{method = pam}}

\item{useProportions}{use proportions or quantities, see details}

\item{pigmentNames}{names for resulting pigment proportions or quantities}

\item{apply.fun}{function passed to \code{aqp::profileApply(APPLY.FUN)} argument, can be used to add progress bars via \code{pbapply::pblapply}, or parallel processing with \code{furrr::future_map}}
}
\value{
For the \code{colorBucket} method, a \code{data.frame} object containing:
\itemize{
\item id column: set according to \code{idname(spc)}
\item \code{.white.pigment}: proportion or quantity of CIE LAB L-values
\item \code{.red.pigment}: proportion or quantity of CIE LAB positive A-values
\item \code{.green.pigment}: proportion or quantity of CIE LAB negative A-values
\item \code{.yellow.pigment}: proportion or quantity of CIE LAB positive B-values
\item \code{.blue.pigment}: proportion or quantity of CIE LAB negative B-values
}

Column names can be adjusted with the \code{pigmentNames} argument.

For the \code{depthSlices} method ...

For the \code{pam} method ...
}
\description{
Generate a color signature for each soil profile in a collection.
}
\details{
See the \href{http://ncss-tech.github.io/AQP/aqp/soil-color-signatures.html}{related tutorial}.
}
\examples{

# trivial example, not very interesting
data(sp1)
depths(sp1) <- id ~ top + bottom

# Munsell notation
sp1$m <- sprintf("\%s \%s/\%s", sp1$hue, sp1$value, sp1$chroma)

# extract color signature
pig <- soilColorSignature(sp1, color = 'm')

}
\references{
https://en.wikipedia.org/wiki/Lab_color_space
}
\seealso{
\code{\link{munsell2rgb}}
}
\author{
D.E. Beaudette
}
