% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-spatial.R
\name{proj4string,SoilProfileCollection-method}
\alias{proj4string,SoilProfileCollection-method}
\alias{wkt,SoilProfileCollection-method}
\alias{proj4string<-,SoilProfileCollection,ANY-method}
\title{Get or Set Coordinate Reference System for SoilProfileCollection}
\usage{
\S4method{proj4string}{SoilProfileCollection}(obj)

\S4method{wkt}{SoilProfileCollection}(obj)

\S4method{proj4string}{SoilProfileCollection,ANY}(obj) <- value
}
\arguments{
\item{obj}{A SoilProfileCollection}

\item{value}{A proj4string}
}
\description{
\code{proj4string()}: Get Coordinate Reference System as PROJ4 String

\verb{wkt():} Get Coordinate Reference System for as Well-Known Text

\verb{proj4string()<-}: Set Coordinate Reference System string for the SoilProfileCollection
}
