% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-level-setters.R
\name{SoilTextureLevels}
\alias{SoilTextureLevels}
\title{Generate a vector of USDA soil texture codes or class names, sorted according to approximate particle size}
\usage{
SoilTextureLevels(which = "codes")
}
\arguments{
\item{which}{'codes' (texture codes) or 'names' (texture class names)}
}
\value{
an ordered factor
}
\description{
Generate a vector of USDA soil texture codes or class names, sorted according to approximate particle size
}
\examples{

# class codes
SoilTextureLevels()

# class names
SoilTextureLevels(which = 'names')
 
}
\references{
\href{https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/research/guide/?cid=nrcs142p2_054184}{Field Book for Describing and Sampling Soils, version 3.0}
}
