% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2Munsell.R
\name{spec2Munsell}
\alias{spec2Munsell}
\title{Convert reflectance spectra to closest Munsell chip}
\usage{
spec2Munsell(x, ...)
}
\arguments{
\item{x}{reflectance spectra, (380nm to 730nm, 10nm resolution)}

\item{...}{further arguments to \code{rgb2Munsell}}
}
\value{
output from \code{rgb2Munsell}
}
\description{
Convert reflectance spectra to closest Munsell chip
}
\note{
D65 illuminant spectra and CIE1931 color matching functions derived from CIE reference data.
}
\references{
Marcus, R.T. (1998). The Measurement of Color. In K. Nassau (Ed.), Color for Science, Art, and Technology (pp. 32-96). North-Holland.

"Selected colorimetric tables in Excel" http://files.cie.co.at/204.xls
}
