% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{spectral.reference}
\alias{spectral.reference}
\title{D65 standard illuminant and CIE1931 color matching functions}
\format{
An object of class \code{data.frame} with 71 rows and 5 columns.
}
\usage{
data(spectral.reference)
}
\description{
D65 standard illuminant and CIE1931 color matching functions
}
\references{
Marcus, R.T. (1998). The Measurement of Color. In K. Nassau (Ed.), Color for Science, Art, and Technology (pp. 32-96). North-Holland.

"Selected colorimetric tables in Excel" http://files.cie.co.at/204.xls
}
\keyword{datasets}
