% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{spectral.reference}
\alias{spectral.reference}
\title{D65 standard illuminant and CIE1931 color matching functions}
\format{
An object of class \code{data.frame} with 71 rows and 5 columns.
}
\usage{
data(spectral.reference)
}
\description{
D65 standard illuminant and CIE1931 color matching functions
}
\references{
Wyszecki, G., & Stiles, W. S., 1982 Color Science: concepts and methods, quantitative data and formulae (2nd ed.). New York: Wiley.

"Selected colorimetric tables in Excel" http://files.cie.co.at/204.xls
}
\keyword{datasets}
