% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\name{subApply}
\alias{subApply}
\alias{subApply,SoilProfileCollection-method}
\title{Subset SPC based on result of performing function on each profile}
\arguments{
\item{object}{A SoilProfileCollection}

\item{.fun, }{A function that takes a single profile, returns \emph{logical} of length 1.}

\item{...}{Additional arguments are passed to \code{.fun}}
}
\value{
A SoilProfileCollection.
}
\description{
\code{subApply()} is a function used for subsetting SoilProfileCollections. It currently does NOT support for "tidy" lexical features in the \code{...} arguments passed to \code{profileApply()}. The expectation is that the function \code{.fun} takes a single-profile SoilProfileCollection and returns a logical value of length one. The use case would be for any logical comparisons that cannot be evaluated inline by \code{subSPC()} because they require more than simple logical operations.
}
\author{
Andrew G. Brown.
}
