% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize,SoilProfileCollection-method}
\title{Summarize a SPC with a set of functions}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{A set of comma-delimited R expressions that resolve to a summary value; may be named. e.g \code{mean = mean(clay, na.rm=TRUE)}}
}
\value{
A SoilProfileCollection.
}
\description{
\code{summarize()} is a function used for summarizing SoilProfileCollections.
}
\author{
Andrew G. Brown.
}
