% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize,SoilProfileCollection-method}
\title{Reduce multiple site or horizon values to a single site value}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{A SoilProfileCollection}

\item{...}{A set of comma-delimited R expressions that resolve to a summary value; may be named. e.g \code{mean = mean(clay, na.rm=TRUE)}}
}
\value{
A SoilProfileCollection
}
\description{
\code{summarize()} is a function used for summarizing SoilProfileCollections. Each summary expression is applied to individual profiles. In the future, higher tiers of grouping, based on one or more variables, will be possible.
}
\author{
Andrew G. Brown.
}
