% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcl_to_classlimit.R
\name{texcl_to_classlimit}
\alias{texcl_to_classlimit}
\title{Convert Texture Class to Class Limits}
\usage{
texcl_to_classlimit(x)
}
\arguments{
\item{x}{\emph{character} or \emph{list}. A vector of texture class codes (e.g. \code{"l"} for "loam", \code{"sicl"} for silty clay loam) without texture class modifiers. If the input is a list, multiple texture classes within each list element are aggregated to create combined class limits.}
}
\value{
A \emph{data.frame} with column names "texcl", "clay_l", "clay_m", "clay_h", "sand_l", "sand_m", "sand_h", "silt_l", "silt_m", "silt_h"
}
\description{
Convert Texture Class to Class Limits
}
\details{
Logic derived from NASIS validation NASIS calculation "Textural Class versus Particle Size Separates" written by Cathy Seybold (last updated 4/07/14)
}
\examples{

texcl_to_classlimit(c("l", "sicl", "cl"))

texcl_to_classlimit(list(c("l", "sicl", "cl")))

}
