% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{union,list,missing-method}
\alias{union,list,missing-method}
\title{[DEPRECATED] union}
\usage{
\S4method{union}{list,missing}(x, y)
}
\arguments{
\item{x}{a list of SoilProfileCollection objects}

\item{y}{Necessary for proper S4 dispatch.}
}
\value{
a SoilProfileCollection
}
\description{
[DEPRECATED] union
}
\examples{

# deprecated, see aqp::combine and aqp::pbindlist

}
\seealso{
\link{pbindlist}
}
