% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{us.state.soils}
\alias{us.state.soils}
\title{US State Soils}
\format{
\describe{
  \item{state}{state name}
  \item{series}{soil series name}
}
}
\usage{
us.state.soils
}
\description{
A listing of the 52 US state soils, including Puerto Rico and Virgin Islands.
}
\keyword{datasets}
