% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilProfileCollection-methods.R
\docType{methods}
\name{validSpatialData,SoilProfileCollection-method}
\alias{validSpatialData,SoilProfileCollection-method}
\alias{validSpatialData}
\title{Check for valid spatial reference of profiles}
\usage{
\S4method{validSpatialData}{SoilProfileCollection}(object)
}
\arguments{
\item{object}{a SoilProfileCollection}
}
\value{
logical \code{TRUE} if column names are defined and correspond to existing data
}
\description{
Are coordinate column names defined in metadata and existing in the SoilProfileCollection?
}
